==> scripts/step02_R_bookdown/04_iv-stim_dv-outcome.Rmd <==
# beh :: outcome ~ stim {#beh-outcome-stim}

## What is the purpose of this notebook? {.unlisted .unnumbered}

Here, I plot the outcome ratings as a function of stimulus intensity

- Main model: `lmer(outcome_rating ~ stim)`
- Main question: do outcome ratings differ as a function of stimulus intensity? We should expect to see a linear effect of stimulus intensity.
- If there is a main effect of cue on expectation ratings, does this cue effect differ depending on task type?
- IV: stim (high / med / low)
- DV: outcome rating

FIX: plot statistics in random effect plot - what is broken?

::::{.refbox}

- https://stackoverflow.com/questions/29402528/append-data-frames-together-in-a-for-loop/29419402
  ::::

```{r load_libraries_3, message=FALSE, warning=FALSE, include=FALSE, paged.print=FALSE}
library(psych)
library(car)
library(lme4)
library(lmerTest)
library(plyr)
library(tidyr)
library(stringr)
library(ggplot2)
library(png)
library(knitr)
library(TMB)
library(sjPlot)
library(ggpubr)
library(gridExtra)
library(merTools)
library(sjstats) #to get ICC
library(broom)
library(tidyverse)
library(GGally)
library(RCurl)
library(rstanarm)
library(reshape)
library(boot)
library(afex)
library(cowplot)
library(readr)
library(rmarkdown)
library(stringr)
library(ICC)
library(gghalves)
library(cueR)
## library(extraoperators)
##library(JWileymisc)
##library(multilevelTools)
library(ggpubr)
main_dir = dirname(dirname(getwd()))
file.sources = list.files(file.path(main_dir, "scripts/step02_R/utils"),
                          pattern="*.R",
                          full.names=TRUE,
                          ignore.case=TRUE)
sapply(file.sources,source,.GlobalEnv)
```

```{r parameters_3, message=FALSE, warning=FALSE, include=FALSE, paged.print=FALSE}
main_dir = dirname(dirname(getwd()))
print(main_dir)
datadir = file.path(main_dir, 'data', 'beh', 'beh02_preproc')
analysis_dir <- file.path(main_dir, "analysis", "mixedeffect", paste0("model04_iv-stim_dv-outcome"), as.character(Sys.Date()))
dir.create(analysis_dir, showWarnings = FALSE, recursive = TRUE)
```

### Load data across tasks and designate common plot variables {.unlisted .unnumbered}

```{r}
# 1. load data and rename columns ______________________________________________
pvc <- cueR::df_load_pvc_beh(datadir, subject_varkey = "src_subject_id", iv = "param_cue_type", dv = "event02_expect_angle", exclude = "sub-0001")

pvc <- cueR::simple_contrast_beh(pvc)

column_mapping <- c("event04_actual_angle" = "OUTCOME",
                    "event02_expect_angle" = "EXPECT",
                    "src_subject_id" = "sub")
pvc <- cueR::df_rename_columns(pvc, column_mapping)
# dropping NA values
pvc <- pvc[!is.na(pvc$OUTCOME), ]

```

```{r}
combined_summary <- data.frame()
```

## Pain

### For the pain task, what is the effect of stimulus intensity on outcome ratings? {.unlisted .unnumbered}

> Stimulus intensity effect is significant. What's also evident is that there is wide variablity in terms of indviduals differences
> The linear effect is reassuring and is similar to that of C-fiber profiles from

```{r echo=FALSE, message=FALSE, warning=FALSE}
taskname <- 'pain'


# 0. Summary and plot related keywords _________________________________________
subject <- "sub";
iv <- "stim_ordered";              dv <- "OUTCOME"
sub_mean <- "mean_per_sub";       group_mean <- "mean_per_sub_norm_mean"
se <- "se"
iv_keyword <- "stim";              dv_keyword <- "outcome"
ggtitle_phrase <- " - Outcome Rating"
ggtitle <- paste0(tools::toTitleCase(taskname), ggtitle_phrase, " (N = ", length(unique(pvc$sub)), ")");
title <- paste0(tools::toTitleCase(taskname), " - ", tools::toTitleCase(dv_keyword))
xlab <- "Stimulus Intensity type";               ylab <- "Outcome Ratings (degree)"
ylim <- c(-10,190);               w = 5; h = 3;
plot_savefname <- file.path(analysis_dir,
                            paste0("raincloud_task-", taskname,"_iv-", iv_keyword,
                                   "_dv-rating-", dv_keyword,"_", as.character(Sys.Date()), ".png")
                            )
color <- c("#FFA79A", "#CA5C4D", "#941100") #c("#4274AD", "#C5263A")


data <- pvc[pvc$task==taskname, ]


# 1. summary stats _____________________________________________________________
outcome_subjectwise <- meanSummary(data, c(subject, iv), dv)
outcome_groupwise <- cueR::summarySEwithin(
    data = outcome_subjectwise,
    measurevar = sub_mean, # variable created from above
    withinvars = c(iv), # iv
    idvar = subject
)

outcome_groupwise$task <- taskname
combined_summary <- rbind(combined_summary, outcome_groupwise)


# 2. plot parameters _____________________________________________________________
g <- cueR::plot_cueexpectancy_onefactor(
        outcome_subjectwise, outcome_groupwise, iv,
        sub_mean, group_mean, se, subject, ggtitle, title,
        xlab="", ylab, taskname,
        w, h, dv_keyword, color, plot_savefname,
        expand_x = TRUE,
        xlim = c(NA, 3), ylim = c(-20, 200),
        x_scale_expansion = c(0, 0.5),
        x_hline_position = 3.5,
        x_hline_linetype = "dashed",
        x_hline_nudge_x = 0.1,
        x_hline_textsize = 3,
        legend_factor_levels = c("High stim", "Med stim", "Low stim"),
        legend_factor_colors = c(  "#941100", "#CA5C4D","#FFA79A" ),
        legend_geom_point_size = 4,
        legend_position = c(-0.1, 0.7),
        legend_widths = c(5, 1)
      )
grid::grid.draw(g)


# 3. mixed effects model _______________________________________________________
model.painexpect <- lmer(OUTCOME ~ STIM_con_linear + STIM_con_quad + (STIM_con_linear|sub), data = data)
sjPlot::tab_model(model.painexpect,
              title = paste0(tools::toTitleCase(taskname), ": \nlmer(OUTCOME ~ STIM + (STIM| sub), data = pvc)"),
              CSS = list(css.table = '+font-size: 12;'))


# 4. compute random effects and save for compiling across PVC tasks ____________
rand_savefname <- file.path(
    analysis_dir,
    paste("randeffect_task-", taskname, "_",
        as.character(Sys.Date()), ".csv",
        sep = ""
    ))
randeffect.P <- cueR::compute_randomeffects(model.painexpect, rand_savefname = rand_savefname, taskname,
                                      new_rand_names = c("rand_intercept", "rand_stimlinear"),
                                      new_fix_names =  c("fix_intercept", "fix_stimlinear", "fix_stimquad"))
```

## Vicarious

### For the vicarious task, what is the effect of stimulus intensity on outcome ratings? {.unlisted .unnumbered}

> Stimulus intensity effects are signficant. This is a great manipulation check, as the video of the vicarious pain condition are categorized in the low/med/high based on a combination of patient ratings and observer ratings.

```{r vicarious_iv-stim_dv-outcome, echo=FALSE, message=FALSE, warning=TRUE, paged.print=FALSE}
taskname <- 'vicarious'

# 0. Summary and plot related keywords _________________________________________
subject <- "sub";
iv <- "stim_ordered";              dv <- "OUTCOME"
sub_mean <- "mean_per_sub";       group_mean <- "mean_per_sub_norm_mean"
se <- "se"
iv_keyword <- "stim";              dv_keyword <- "outcome"
ggtitle_phrase <- " - Outcome Rating"
ggtitle <- paste0(tools::toTitleCase(taskname), ggtitle_phrase, " (N = ", length(unique(pvc$sub)), ")");
title <- paste0(tools::toTitleCase(taskname), " - ", tools::toTitleCase(dv_keyword))
xlab <- "Stimulus Intensity type";               ylab <- "Outcome Ratings (degree)"
ylim <- c(-10,190);               w = 5; h = 3;
plot_savefname <- file.path(analysis_dir,
                            paste0("raincloud_task-", taskname,"_iv-", iv_keyword,
                                   "_dv-rating-", dv_keyword,"_", as.character(Sys.Date()), ".png")
                            )
color <- c("#C6FF8A", "#63C76E", "#008F51")


data <- pvc[pvc$task==taskname, ]


# 1. summary stats _____________________________________________________________
outcome_subjectwise <- meanSummary(data, c(subject, iv), dv)
outcome_groupwise <- cueR::summarySEwithin(
    data = outcome_subjectwise,
    measurevar = sub_mean, # variable created from above
    withinvars = c(iv), # iv
    idvar = subject
)

outcome_groupwise$task <- taskname
combined_summary <- rbind(combined_summary, outcome_groupwise)


# 2. plot parameters _____________________________________________________________
g <- cueR::plot_cueexpectancy_onefactor(
        outcome_subjectwise, outcome_groupwise, iv,
        sub_mean, group_mean, se, subject, ggtitle, title,
        xlab="", ylab, taskname,
        w, h, dv_keyword, color, plot_savefname,
        expand_x = TRUE,
        xlim = c(NA, 3), ylim = c(-20, 200),
        x_scale_expansion = c(0, 0.5),
        x_hline_position = 3.5,
        x_hline_linetype = "dashed",
        x_hline_nudge_x = 0.1,
        x_hline_textsize = 3,
        legend_factor_levels = c("High stim", "Med stim", "Low stim"),
        legend_factor_colors = c( "#008F51", "#63C76E", "#C6FF8A" ),
        legend_geom_point_size = 4,
        legend_position = c(-0.1, 0.7),
        legend_widths = c(5, 1)
      )
grid::grid.draw(g)


# 3. mixed effects model _______________________________________________________
model.vicexpect <- lmer(OUTCOME ~ STIM_con_linear + STIM_con_quad + (STIM_con_linear|sub), data = data)
sjPlot::tab_model(model.vicexpect,
              title = paste0(tools::toTitleCase(taskname), ": \nlmer(OUTCOME ~ STIM + (STIM| sub), data = pvc)"),
              CSS = list(css.table = '+font-size: 12;'))


# 4. compute random effects and save for compiling across PVC tasks ____________
rand_savefname <- file.path(
    analysis_dir,
    paste("randeffect_task-", taskname, "_",
        as.character(Sys.Date()), ".csv",
        sep = ""
    ))
randeffect.V <- cueR::compute_randomeffects(model.vicexpect, rand_savefname = rand_savefname, taskname,
                                      new_rand_names = c("rand_intercept", "rand_stimlinear"),
                                      new_fix_names =  c("fix_intercept", "fix_stimlinear", "fix_stimquad"))
```

## Cognitive

### For the cognitive task, what is the effect of stimulus intensity on outcome ratings? {.unlisted .unnumbered}

> Stimulus intensity effects, both linear and quadratic effects, are signfiicant. However, the plots indicate that the medium and high levels are almost similarly processed, except for that one participant that has a linear trend.

```{r cognitive_iv-stim_dv-outcome, echo=FALSE, message=TRUE, warning=TRUE, paged.print=FALSE}

taskname <- 'cognitive'

# 0. Summary and plot related keywords _________________________________________
subject <- "sub";
iv <- "stim_ordered";              dv <- "OUTCOME"
sub_mean <- "mean_per_sub";       group_mean <- "mean_per_sub_norm_mean"
se <- "se"
iv_keyword <- "stim";              dv_keyword <- "outcome"
ggtitle_phrase <- " - Outcome Rating"
ggtitle <- paste0(tools::toTitleCase(taskname), ggtitle_phrase, " (N = ", length(unique(pvc$sub)), ")");
title <- paste0(tools::toTitleCase(taskname), " - ", tools::toTitleCase(dv_keyword))
xlab <- "Stimulus Intensity type";               ylab <- "Outcome Ratings (degree)"
ylim <- c(-10,190);               w = 5; h = 3;
plot_savefname <- file.path(analysis_dir,
                            paste0("raincloud_task-", taskname,"_iv-", iv_keyword,
                                   "_dv-rating-", dv_keyword,"_", as.character(Sys.Date()), ".png")
                            )
color <- c("#8EA1FF", "#485DC8", "#011891")


data <- pvc[pvc$task==taskname, ]


# 1. summary stats _____________________________________________________________
outcome_subjectwise <- meanSummary(data, c(subject, iv), dv)
outcome_groupwise <- cueR::summarySEwithin(
    data = outcome_subjectwise,
    measurevar = sub_mean, # variable created from above
    withinvars = c(iv), # iv
    idvar = subject
)

outcome_groupwise$task <- taskname
combined_summary <- rbind(combined_summary, outcome_groupwise)


# 2. plot parameters _____________________________________________________________
g <- cueR::plot_cueexpectancy_onefactor(
        outcome_subjectwise, outcome_groupwise, iv,
        sub_mean, group_mean, se, subject, ggtitle, title,
        xlab="", ylab, taskname,
        w, h, dv_keyword, color, plot_savefname,
        expand_x = TRUE,
        xlim = c(NA, 3), ylim = c(-20, 200),
        x_scale_expansion = c(0, 0.5),
        x_hline_position = 3.5,
        x_hline_linetype = "dashed",
        x_hline_nudge_x = 0.1,
        x_hline_textsize = 3,
        legend_factor_levels = c("High stim", "Med stim", "Low stim"),
        legend_factor_colors = c(  "#011891","#485DC8","#8EA1FF"),
        legend_geom_point_size = 4,
        legend_position = c(-0.1, 0.7),
        legend_widths = c(5, 1)
      )
grid::grid.draw(g)


# 3. mixed effects model _______________________________________________________
model.cogexpect <- lmer(OUTCOME ~ STIM_con_linear + STIM_con_quad + (STIM_con_linear|sub), data = data)
sjPlot::tab_model(model.cogexpect,
              title = paste0(tools::toTitleCase(taskname), ": \nlmer(OUTCOME ~ STIM + (STIM| sub), data = pvc)"),
              CSS = list(css.table = '+font-size: 12;'))


# 4. compute random effects and save for compiling across PVC tasks ____________
rand_savefname <- file.path(
    analysis_dir,
    paste("randeffect_task-", taskname, "_",
        as.character(Sys.Date()), ".csv",
        sep = ""
    ))
randeffect.C <- cueR::compute_randomeffects(model.cogexpect, rand_savefname = rand_savefname, taskname,
                                      new_rand_names = c("rand_intercept", "rand_stimlinear"),
                                      new_fix_names =  c("fix_intercept", "fix_stimlinear", "fix_stimquad"))
```

## Lineplot

> plotting line plots; identical to raincloud plots above, just different form
> little errors bars are within-subject standard errors.

```{r echo=FALSE}
library(ggpubr)
DATA <- as.data.frame(combined_summary)
color <- c( "#4575B4", "#D73027")
LINEIV1 <- "stim_ordered"
MEAN <- "mean_per_sub_norm_mean"
ERROR <- "se"
dv_keyword = "outcome"


p1 = plot_lineplot_onefactor(DATA, 'pain',
               LINEIV1, MEAN, ERROR, color, xlab = "Stimulus intensity" , ylab= "Outcome rating", ggtitle = 'pain' )
p2 = plot_lineplot_onefactor(DATA,'vicarious',
               LINEIV1, MEAN, ERROR, color,xlab = "Stimulus intensity" , ylab= "Outcome rating",ggtitle = 'vicarious')
p3 = plot_lineplot_onefactor(DATA, 'cognitive',
               LINEIV1, MEAN, ERROR, color,xlab = "Stimulus intensity" , ylab= "Outcome rating",ggtitle = 'cognitive')

ggpubr::ggarrange(p1,p2,p3,ncol = 3, nrow = 1, common.legend = TRUE,legend = "bottom")
# plot_filename = file.path(analysis_dir,
#                           paste('lineplot_task-all_rating-',dv_keyword,'.png', sep = ""))
# ggsave(plot_filename, width = 15, height = 6)
```

## individual differences in outcome rating stim effect

> random effects of stimulus intensity effects across tasks.

```{r echo=FALSE, message=FALSE, warning=FALSE}

# combine randome effects across tasks _________________________________________
pvc.rand <- rbind(randeffect.P, randeffect.V, randeffect.C)
pvc_rand_subset <- subset(pvc.rand, select = c(task, subj, rand_stimlinear))
pvc_rand <- spread(pvc_rand_subset, key = task, value = rand_stimlinear)
# filter dataframe: some pariticpant do not have data in other tasks, thereby they have missing values across PVC tasks
# Assuming df is your dataframe
# Replace col1, col2, col3 with your actual column names
pvc_rand.filter <- pvc_rand %>% filter(!is.na(pain) & !is.na(cognitive) & !is.na(vicarious))


pv <- plot_ggplot_correlation(data = pvc_rand.filter, x = 'vicarious', y = 'pain', p_acc = 0.001, r_acc = 0.01, limit_min = -70, limit_max = 70, label_position = 60)
vc <- plot_ggplot_correlation(data = pvc_rand.filter, x = 'cognitive', y = 'vicarious', p_acc = 0.001, r_acc = 0.01, limit_min = -70, limit_max = 70,  label_position = 100)
cp <- plot_ggplot_correlation(data = pvc_rand.filter, x = 'pain', y = 'cognitive', p_acc = 0.001, r_acc = 0.01, limit_min = -30, limit_max = 30, label_position =60)


# combine plots and add title __________________________________________________
plots <- ggpubr::ggarrange(pv, vc, cp, ncol = 3, nrow = 1, common.legend = FALSE, legend = "bottom")
plots_title <- annotate_figure(plots, top = text_grob(paste0("individual differences\n - ", iv_keyword, " effects from ",dv_keyword," ratings"), color = "black", face = "bold", size = 15))
plots_title


# save plots ___________________________________________________________________
# save_plotname <- file.path(
#     analysis_dir,
#     paste("randeffect_scatterplot_task-all_",
#         as.character(Sys.Date()), ".png",
#         sep = ""
#     )
# )
# ggsave(save_plotname, width = 10, height = 3)

```
