==> scripts/step02_R_bookdown/08_iv-cue-stim_dv-RT_FIX.Rmd <==
# [beh] RT ~ cue \* stim {#ch08_RT-cueXstim}

```
author: "Heejung Jung"
date: '06/27/2022'
output:
  html_document:
    toc: true
    theme: united
    code_folding: hide
editor_options:
  markdown:
    wrap: 72
```

"""
This Rmarkdown tests the cue effect (high vs. low) on Reaction time and performance in the cognitive, mental-rotation tasks. We also test for stimulus intensity effects.
"""

## Overview model 08 iv-cue dv-RT summary

- left = diff, right = same
- model 1: Does RT differ as a function of cue type and stimulus intensity?
- model 1-1: Does RT differ as a function of cue, ONLY for the correct trials?
- model 1-2: Does RT differ as a function of cue, ONLY for the incorrect trials?
- model 2: would log-transforming help? Do we see a cue effect on the log-transformmed RTs?

```{block2, type='rmdnote'}
TODO: 06/28/2022
* change variable names and two factor code accordingly
```

```{r load_libraries_7, message=TRUE, warning=TRUE, include=FALSE, paged.print=TRUE, results='hide'}
library(psych)
library(car)
library(lme4)
library(lmerTest)
library(plyr)
library(tidyr)
library(stringr)
library(ggplot2)
library(png)
library(knitr)
library(TMB)
library(sjPlot)
library(ggpubr)
library(gridExtra)
library(merTools)
library(sjstats) #to get ICC
library(broom)
library(tidyverse)
library(GGally)
library(RCurl)
library(rstanarm)
library(reshape)
library(boot)
library(afex)
library(cowplot)
library(readr)
library(lavaan)
library(rmarkdown)
library(readr)
library(caTools)
library(bitops)
library(stringr)
library(ggpubr)
#library(extraoperators)
#library(JWileymisc)
#library(multilevelTools)
#library(Hmisc)
#library(Rmisc)
#library(dplyr)
library(gghalves)

library(r2mlm)
main_dir = dirname(dirname(getwd()))
file.sources = list.files(file.path(main_dir, "scripts/step02_R/utils"),
                          pattern="*.R",
                          full.names=TRUE,
                          ignore.case=TRUE)
sapply(file.sources,source,.GlobalEnv)
```

```{r main_directory_7}
main_dir = dirname(dirname(getwd()))
print(main_dir)
datadir = file.path(main_dir, 'data', 'beh', 'beh02_preproc')
```

## Prepare data and preprocess

### 1) load data {.unlisted .unnumbered}

```{r load_data_7, include=FALSE}
# parameters _____________________________________ # nolint
subject_varkey <- "src_subject_id"
iv <- "param_cue_type"
dv <- "event03_RT"
dv_keyword <- "RT"
xlab <- ""
taskname <- "cognitive"
ylab <- "ratings (degree)"
subject <- "subject"
exclude <- "sub-0999|sub-0001"
data <- df_load_beh(datadir, taskname = taskname,
                            subject_varkey = subject_varkey,
                            iv = iv, exclude = exclude)
data$event03_RT <- data$event03_stimulusC_reseponseonset - data$event03_stimulus_displayonset
```

### 2) plot RT distribution per participant {.unlisted .unnumbered}

- the purpose is to identify whether RTs are distributed from 0-5 sec or not
- From this, we can also identify the quality of the data and determine if we need to scrub or remove participants.

```{r plot_RT_distribution, echo=TRUE, fig.width = 10, fig.height=12}
ggplot(data, aes(x = event03_RT, group = subject)) +
  geom_histogram(
    color = "darkblue",
    fill = "lightblue",
    binwidth = 0.25,
    bins = 20
  ) +
  facet_wrap( ~ subject, ncol = 10) +
  theme(axis.text = element_text(size = 7), text = element_text(size = 12)) +
  xlab("Reaction time for mental rotation task")
```

- Some participants may not have responded within time limit.
- The button may not have registered the correct onset
- I may have to remove the RTs with 5 sec.

---

### 3) exclude participants with RT of 5 seconds {.unlisted .unnumbered}

```{r include=FALSE}
# parameters _____________________________________ # nolint
subject_varkey <- "src_subject_id"
iv <- "param_cue_type"
dv <- "event03_RT"
dv_keyword <- "RT"
xlab <- ""
taskname <- "cognitive"

ylab <- "ratings (degree)"
subject <- "subject"
exclude <- "sub-0999|sub-0001|sub-0002|sub-0003|sub-0004|sub-0005|sub-0006|sub-0007|sub-0008|sub-0009|sub-0010|sub-0011"

# load data _____________________________________
data <-  df_load_beh(datadir, taskname = taskname, subject_varkey = subject_varkey, iv = iv, exclude = exclude)
data$event03_RT <- data$event03_stimulusC_reseponseonset - data$event03_stimulus_displayonset

analysis_dir <- file.path(main_dir, "analysis", "mixedeffect", "model08_iv-cue-stim_dv-RT", as.character(Sys.Date()))
dir.create(analysis_dir, showWarnings = FALSE, recursive = TRUE)

data$event03_response_samediff <- mapvalues(data$event03_stimulusC_response,
                                                from = c(1, 2),
                                                to = c("diff", "same"))

data$event03_correct <- ifelse(data$event03_C_stim_match == data$event03_response_samediff, 1, ifelse(data$event03_C_stim_match != data$event03_response_samediff, 0, "NA"))
```

```{r, fig.width = 10, fig.height=12}
ggplot(data, aes(x = event03_RT, group = subject)) +
  geom_histogram(
    color = "darkblue",
    fill = "lightblue",
    binwidth = 0.25,
    bins = 20
  ) +
  facet_wrap( ~ subject, ncol = 10) +
  theme(axis.text = element_text(size = 7), text = element_text(size = 12)) +
  xlab("Reaction time for mental rotation task")
```

## model 1: Does RT differ as a function of cue type and stimulus intensity?

- IV: cue (high vs. low)
- DV: RT of the incorrect trials
- contrast code two factors -- stimulus intensity and cue type
- plotting all trials (including correct and incorrect trials)

```{r include=FALSE}
# contrast code 1 linear _____________________________________
data$stim_con_linear[data$param_stimulus_type == "low_stim"] <- -0.5
data$stim_con_linear[data$param_stimulus_type == "med_stim"] <- 0
data$stim_con_linear[data$param_stimulus_type == "high_stim"] <- 0.5

# contrast code 2 quadratic _____________________________________
data$stim_con_quad[data$param_stimulus_type == "low_stim"] <- -0.33
data$stim_con_quad[data$param_stimulus_type == "med_stim"] <- 0.66
data$stim_con_quad[data$param_stimulus_type == "high_stim"] <- -0.33

# social cue contrast _____________________________________
# data$social_cue[data$param_cue_type == "low_cue"] <- -0.5 # social influence task
# data$social_cue[data$param_cue_type == "high_cue"] <- 0.5 # no influence task
# data$cue_factor = factor(data$social_cue)

data$social_cue[data$param_cue_type == 'low_cue'] <- -0.5 # social influence task
data$social_cue[data$param_cue_type == 'high_cue'] <- 0.5 # no influence task
data$cue_factor = factor(data$param_cue_type)

# factorize group variable _____________________________________
data$subject = factor(data$src_subject_id)
```

```{r echo=FALSE}

# ------------------------------------------------------------------------------
#                 task parameters
# ------------------------------------------------------------------------------
subject_varkey <- "src_subject_id"
iv <- "param_cue_type"
stim_con1 = "stim_con_linear"
stim_con2 = "stim_con_quad"
dv <- "event03_RT"
iv_keyword <- "cue-stim"
dv_keyword <- "RT"
xlab <- ""
taskname <- "cognitive"
ylim = c(0,5)

# lmer filename ________________________________________________________________
model_savefname <- file.path(
        analysis_dir,
        paste("lmer_task-", taskname, "_iv-", iv_keyword,
            "_dv-", dv_keyword,
            "_", as.character(Sys.Date()), ".txt",
            sep = ""
        )
    )

# removing NA values ___________________________________________________________
data_clean = data[!is.na(data$event03_RT),]
data_clean$subject = factor(data_clean$src_subject_id)

# ------------------------------------------------------------------------------
#                 lmer modeling
# ------------------------------------------------------------------------------
cooksd <- lmer_twofactor_cooksd(data = data_clean,
                                     taskname = "cognitive",
                                     iv = "cue_factor",
                                     stim_con1 = "stim_con_linear",
                                     stim_con2 = "stim_con_quad",
                                     dv = "event03_RT",
                                     subject_keyword = "subject",
                                     dv_keyword = "RT",
                                     model_savefname = model_savefname,
                                     effects = "random_intercept",
                                     print_lmer_output = FALSE)
influential <- as.numeric(names(cooksd)[
    (cooksd > (4 / as.numeric(length(unique(data_clean$subject)))))])
data_screen <- data_clean[-influential, ]

# reordering for plots _________________________________________________________
data_clean$cue_name[data_clean$param_cue_type == "high_cue"] <- "high cue"
data_clean$cue_name[data_clean$param_cue_type == "low_cue"] <- "low cue" # no influence task

data_clean$stim_name[data_clean$param_stimulus_type == "high_stim"] <- "high" # no influence task
data_clean$stim_name[data_clean$param_stimulus_type == "med_stim"] <- "med" # no influence task
data_clean$stim_name[data_clean$param_stimulus_type == "low_stim"] <- "low" # no influence task

data_clean$stim_ordered <- factor(data_clean$stim_name, levels=c("low", "med", "high"))
data_clean$cue_ordered <- factor(data_clean$cue_name, levels=c("low cue", "high cue"))
model_iv1 = "stim_ordered";model_iv2 = "cue_ordered"


# ------------------------------------------------------------------------------
#                 summary statistics
# ------------------------------------------------------------------------------
subjectwise <- meanSummary(data_clean, c(subject, model_iv1, model_iv2), dv)
groupwise <- summarySEwithin(
        data = subjectwise,
        measurevar = "mean_per_sub", # variable created from above
        withinvars = c(model_iv1, model_iv2), # iv
        idvar = "subject"
    )


# ------------------------------------------------------------------------------
#                 plot parameters
# ------------------------------------------------------------------------------
subjectwise_mean <- "mean_per_sub";    group_mean <- "mean_per_sub_norm_mean"
se <- "se";    subject <- "subject"
ggtitle <- paste(taskname, " - Reaction Time (s)");    title <- paste(taskname, " - RT")
xlab <- "";    ylab <- "Reaction Time (s)";
w = 5; h = 3; dv_keyword <- "RT"
if (any(startsWith(dv_keyword, c("expect", "Expect")))) {
        color <- c("#1B9E77", "#D95F02")
    } else {
        color <- c("#4274AD", "#C5263A")
    }
plot_savefname <- file.path(
        analysis_dir,
        paste("raincloud_task-", taskname,
              "_iv-", iv_keyword,"_dv-", dv_keyword,
            "_", as.character(Sys.Date()), ".png",
            sep = ""
        )
    )
plot_halfrainclouds_twofactor(
        subjectwise, groupwise,
        model_iv1, model_iv2, subjectwise_mean, group_mean, se, subject,
        ggtitle, title, xlab, ylab, task_name, ylim,
        w, h, dv_keyword, color, plot_savefname
    )


# ------------------------------------------------------------------------------
#                 random effects
# ------------------------------------------------------------------------------
# save random effects for individual difference analysis _______________________
randEffect$newcoef <- mapvalues(randEffect$term,
    from = c("(Intercept)"),
    to = c("rand_intercept")
)
rand_subset <- subset(randEffect, select = -c(grpvar, term, condsd))
wide_rand <- spread(rand_subset, key = newcoef, value = condval)

wide_fix <- do.call(
    "rbind",
    replicate(nrow(wide_rand),
        as.data.frame(t(as.matrix(fixEffect))),
        simplify = FALSE
    )
)
rownames(wide_fix) <- NULL
new_wide_fix <- dplyr::rename(wide_fix,
    fix_intercept = `(Intercept)`,
    fix_cue = `cue_factorlow_cue`,
)

total <- cbind(wide_rand, new_wide_fix)
total$task <- taskname
new_total <- total %>% dplyr::select(task, everything())
new_total <- dplyr::rename(total, subj = grp)

rand_savefname <- file.path(
    analysis_dir,
    paste("randeffect_task-", taskname,
          "_iv-", iv_keyword,"_dv-", dv_keyword,
        "_",as.character(Sys.Date()), "_outlier-cooksd.csv",
        sep = ""
    )
)
write.csv(new_total, rand_savefname, row.names = FALSE)
```

```
[1] "model:  Rt  ratings -  cognitive"
event03_RT ~ cue_factor * stim_con_linear + cue_factor * stim_con_quad +
    (1 | subject)
<environment: 0x7f8725548c80>
Linear mixed model fit by REML. t-tests use Satterthwaite's method ['lmerModLmerTest']
Formula: as.formula(model_string)
   Data: data

REML criterion at convergence: 14116

Scaled residuals:
    Min      1Q  Median      3Q     Max
-4.7775 -0.7079 -0.0816  0.6722  3.3218

Random effects:
 Groups   Name        Variance Std.Dev.
 subject  (Intercept) 0.1654   0.4066
 Residual             0.5433   0.7371
Number of obs: 6189, groups:  subject, 105

Fixed effects:
                                    Estimate Std. Error         df t value Pr(>|t|)
(Intercept)                          3.16931    0.04202  115.33041  75.430  < 2e-16 ***
cue_factorlow_cue                   -0.02908    0.01875 6080.59279  -1.551   0.1210
stim_con_linear                      0.19877    0.03240 6080.29673   6.134  9.1e-10 ***
stim_con_quad                        0.05344    0.02842 6080.64300   1.880   0.0602 .
cue_factorlow_cue:stim_con_linear    0.07954    0.04587 6080.29621   1.734   0.0829 .
cue_factorlow_cue:stim_con_quad     -0.03609    0.04025 6080.52154  -0.897   0.3700
---
Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

Correlation of Fixed Effects:
                    (Intr) c_fct_ stm_cn_l stm_cn_q c_fctrlw_c:stm_cn_l
cu_fctrlw_c         -0.223
stim_cn_lnr          0.005 -0.012
stim_con_qd          0.001 -0.002 -0.012
c_fctrlw_c:stm_cn_l -0.004  0.020 -0.706    0.008
c_fctrlw_c:stm_cn_q -0.001  0.004  0.008   -0.706   -0.014
event03_RT ~ cue_factor * stim_con_linear + cue_factor * stim_con_quad +
    (1 | subject)
<environment: 0x7f8725548c80>

```

```{block2, type='rmdnote'}
Model 1 Interim summary
* Research question: Does RT differ as a function of high vs. low cue across stimulus intensity?
* Conclusion: linear effect of intensity is significant.
* Next step: It's interesting that there is no cue effect on RT, while the cue effect on ratings are robust. We see if this effect is potentially present in only certain type of trials, namely the correct vs. incorrect trials of the mental rotation task
```

## model 1-1: Does RT differ as a function of cue, ONLY for the correct trials?

- IV:
  - cue (high vs. low)
  - stimulus intensity (high vs. medium vs. low)
- DV: RT of the correct trials
- Subsetting trials (identical to model 1, except for subsetting correct trials)

```{r include=FALSE}
# parameters ___________________________________________________________________
subject_varkey <- "src_subject_id"
iv <- "param_cue_type"
dv <- "event03_RT"
dv_keyword <- "RT-correct"
xlab <- ""
taskname <- "cognitive"
ylim = c(0,5)

# lmer filename ________________________________________________________________
model_savefname <- file.path(
        analysis_dir,
        paste("lmer_task-", taskname,"_iv-", iv_keyword,
            "_dv-", dv_keyword,
            "_", as.character(Sys.Date()), ".txt",
            sep = ""
        )
    )

# removing NA values ___________________________________________________________

data_clean = data[!is.na(data$event03_correct),]
data_c = data[data_clean$event03_correct == 1,]
data_c$subject = factor(data_c$src_subject_id)


# lmer model ___________________________________________________________________
model_onefactor_correct <- lmer(event03_RT ~ cue_factor*stim_con_linear + cue_factor*stim_con_quad + (1 | src_subject_id), data = data_c)
fixEffect <- as.data.frame(fixef(model_onefactor_correct))
randEffect <- as.data.frame(ranef(model_onefactor_correct))
cooksd <- cooks.distance(model_onefactor_correct)
influential <- as.numeric(names(cooksd)[
    (cooksd > (4 / as.numeric(length(unique(data_c$subject)))))])
data_screen <- data_c[-influential, ]
#equatiomatic::extract_eq(model_onefactor_correct)



# summary statistics for plots _________________________________________________
subjectwise <- meanSummary(data_screen, c(subject, iv), dv)
groupwise <- summarySEwithin(
        data = subjectwise,
        measurevar = "mean_per_sub", # variable created from above
        withinvars = c(iv), # iv
        idvar = "subject"
    )

subjectwise_mean <- "mean_per_sub";    group_mean <- "mean_per_sub_norm_mean"
se <- "se";    subject <- "subject"
ggtitle <- paste(taskname, dv_keyword);    title <- paste(taskname, " - RT")
xlab <- "";    ylab <- "Reaction Time (s)";
w = 5; h = 3;
if (any(startsWith(dv_keyword, c("expect", "Expect")))) {
        color <- c("#1B9E77", "#D95F02")
    } else {
        color <- c("#4274AD", "#C5263A")
    }
plot_savefname <- file.path(
        analysis_dir,
        paste("raincloud_task-", taskname,
            "_iv-", iv_keyword,"_dv-", dv_keyword,
            "_", as.character(Sys.Date()), ".png",
            sep = ""
        )
    )
plot_halfrainclouds_onefactor(
        subjectwise, groupwise,
        iv, subjectwise_mean, group_mean, se, subject,
        ggtitle, title, xlab, ylab, task_name, ylim,
        w, h, dv_keyword, color, plot_savefname
    )

```

```{block2, type='rmdnote'}
Model 1-1 Interim summary (correct trials)

* Research question: Only using correct trials, does RT differ as a function of high vs. low cue?
* Conclusion: No. Even within the subset of correct trials, RT does not differ as a function of cue.
* Next step: Is there a cue effect on RT, only for the incorrect trials? Perhaps the cue had an effect on one's expectation, and the mismatch of thee cues led to incorrect performance, reflected in the RTs
```

## model 1-2: Does RT differ as a function of cue, ONLY for the incorrect trials?

- IV: cue (high vs. low)
- DV: RT of the incorrect trials
- Subsetting trials (identical to model 1, except for subsetting incorrect trials)

```{r include=FALSE}
# parameters ___________________________________________________________________
subject_varkey <- "src_subject_id"
iv <- "param_cue_type"
dv <- "event03_RT"
dv_keyword <- "RT-incorrect"
xlab <- ""
taskname <- "cognitive"
ylim = c(0,5)

# lmer filename ________________________________________________________________
model_savefname <- file.path(
        analysis_dir,
        paste("lmer_task-", taskname,"_iv-", iv_keyword,
            "_dv-", dv_keyword,
            "_", as.character(Sys.Date()), ".txt",
            sep = ""
        )
    )

# removing NA values ___________________________________________________________

data_clean = data[!is.na(data$event03_correct),]
data_i = data[data_clean$event03_correct == 0,]
data_i$subject = factor(data_i$src_subject_id)

# lmer model ___________________________________________________________________
model_onefactor_incorrect <- lmer(event03_RT ~ cue_factor*stim_con_linear + cue_factor*stim_con_quad + (1 | src_subject_id), data = data_i)
fixEffect <- as.data.frame(fixef(model_onefactor_incorrect))
randEffect <- as.data.frame(ranef(model_onefactor_incorrect))
cooksd <- cooks.distance(model_onefactor_incorrect)
influential <- as.numeric(names(cooksd)[
    (cooksd > (4 / as.numeric(length(unique(data_i$subject)))))])
data_screen_i <- data_i[-influential, ]
#equatiomatic::extract_eq(model_onefactor_incorrect)

# summary statistics for plots _________________________________________________
subjectwise <- meanSummary(data_screen_i, c(subject, iv), dv)
groupwise <- summarySEwithin(
        data = subjectwise,
        measurevar = "mean_per_sub", # variable created from above
        withinvars = c(iv), # iv
        idvar = "subject"
    )

subjectwise_mean <- "mean_per_sub";    group_mean <- "mean_per_sub_norm_mean"
se <- "se";    subject <- "subject"
ggtitle <- paste(taskname, dv_keyword);    title <- paste(taskname, " - RT")
xlab <- "";    ylab <- "Reaction Time (s)";
w = 5; h = 3;
if (any(startsWith(dv_keyword, c("expect", "Expect")))) {
        color <- c("#1B9E77", "#D95F02")
    } else {
        color <- c("#4274AD", "#C5263A")
    }
plot_savefname <- file.path(
        analysis_dir,
        paste("raincloud_task-", taskname,
            "_iv-", iv_keyword,"_dv-", dv_keyword,
            "_", as.character(Sys.Date()), ".png",
            sep = ""
        )
    )
plot_halfrainclouds_onefactor(
        subjectwise, groupwise,
        iv, subjectwise_mean, group_mean, se, subject,
        ggtitle, title, xlab, ylab, task_name, ylim,
        w, h, dv_keyword, color, plot_savefname
    )
```

```{block2, type='rmdnote'}
Model 1-2 Interim summary (incorrect trials)
* Research question: Only using incorrect trials, does RT differ as a function of high vs. low cue?
* Conclusion: No. Even within the subset of incorrect trials, RT does not differ as a function of cue
* Next step: Reaction times are data points that are non-normal; mostly a skewed distribution. In order to deal with this skewness, we log-transform the data and investigate the same question, whether the cues have a signficant effect on reaction times.
```

## model 2: Log transformation - Do we see a cue effect on the log-transformmed RTs?

- IV: cue (high vs. low)
- DV: log-transformmed RT

```{r model_2_cue_effect_on_log-RT_7, message=TRUE, warning=TRUE, paged.print=FALSE}
# parameters ___________________________________________________________________
subject_varkey <- "src_subject_id"
iv <- "cue_factor"
dv <- "log_RT"
dv_keyword <- "RT-log"
xlab <- ""
taskname <- "cognitive"
ylim = c(0,5)

# lmer filename ________________________________________________________________
model_savefname <- file.path(
        analysis_dir,
        paste("lmer_task-", taskname,"_iv-", iv_keyword,
            "_dv-", dv_keyword, as.character(Sys.Date()), ".txt",
            sep = ""
        )
    )

# removing NA values ___________________________________________________________
data_clean = data[!is.na(data$event03_correct),]
data_clean$log_RT = log(data_clean$event03_RT)
data_clean$subject = factor(data_clean$src_subject_id)

# lmer model ___________________________________________________________________
model_onefactor_log <- lmer(log_RT ~ cue_factor + (1 | subject), data = data_clean)
summary(model_onefactor_log)

fixEffect <- as.data.frame(fixef(model_onefactor_log))
randEffect <- as.data.frame(ranef(model_onefactor_log))
cooksd <- cooks.distance(model_onefactor_log)
influential <- as.numeric(names(cooksd)[
    (cooksd > (4 / as.numeric(length(unique(data_clean$subject)))))])
data_screen_log <- data_clean[-influential, ]
#equatiomatic::extract_eq(model_onefactor_log)

# summary statistics for plots _________________________________________________
subjectwise <- meanSummary(data_screen_log, c(subject, iv), dv)
groupwise <- summarySEwithin(
        data = subjectwise,
        measurevar = "mean_per_sub", # variable created from above
        withinvars = c(iv), # iv
        idvar = "subject"
    )

subjectwise_mean <- "mean_per_sub";    group_mean <- "mean_per_sub_norm_mean"
se <- "se";
ggtitle <- paste(taskname, dv_keyword);    title <- paste(taskname, " - RT")
xlab <- "";    ylab <- "Reaction Time (s)";
w = 5; h = 3;
if (any(startsWith(dv_keyword, c("expect", "Expect")))) {
        color <- c("#1B9E77", "#D95F02")
    } else {
        color <- c("#4274AD", "#C5263A")
    }
plot_savefname <- file.path(
        analysis_dir,
        paste("raincloud_task-", taskname,
            "_iv-", iv_keyword,"_dv-", dv_keyword,"_",
            as.character(Sys.Date()), ".png",
            sep = ""
        )
    )
plot_halfrainclouds_onefactor(
        subjectwise, groupwise,
        iv, subjectwise_mean, group_mean, se, subject,
        ggtitle, title, xlab, ylab, task_name, ylim,
        w, h, dv_keyword, color, plot_savefname
    )

# save random effects for individual difference analysis _______________________
fixEffect <- as.data.frame(fixef(model_onefactor_log))
randEffect <- as.data.frame(ranef(model_onefactor_log))
randEffect$newcoef <- mapvalues(randEffect$term,
    from = c("(Intercept)"),
    to = c("rand_intercept")
)

rand_subset <- subset(randEffect, select = -c(grpvar, term, condsd))
wide_rand <- spread(rand_subset, key = newcoef, value = condval)

wide_fix <- do.call(
    "rbind",
    replicate(nrow(wide_rand),
        as.data.frame(t(as.matrix(fixEffect))),
        simplify = FALSE
    )
)
rownames(wide_fix) <- NULL
new_wide_fix <- dplyr::rename(wide_fix,
    fix_intercept = `(Intercept)`,
    fix_cue = `cue_factorlow_cue`,
)

total <- cbind(wide_rand, new_wide_fix)
total$task <- taskname
new_total <- total %>% dplyr::select(task, everything())
new_total <- dplyr::rename(total, subj = grp)

rand_savefname <- file.path(
    analysis_dir,
    paste("randeffect_task-", taskname,"_iv-", iv_keyword,"_dv-", dv_keyword,
        "_outlier-cooksd_", as.character(Sys.Date()), ".csv",
        sep = ""
    )
)


write.csv(new_total, rand_savefname, row.names = FALSE)
```

```{block2, type='rmdnote'}
* Research question: Does log transformming help? After log-transformming RT, does RT differ as a function of high vs. low cue?
* Conclusion 2: No, log transformmed or not, there is no significant cue effect on RT
```

## Conclusion across model 1 and 2

- There is no substansive evidence that suggeest cue effects on reaction time. In other words, the time to respond to a rotated figure does not differ as a function of high or low cue.
