==> scripts/step02_R_bookdown/41_painpathway.Rmd <==
# fMRI :: ROI ~ cue \* stim {#painpathway}

## What is the purpose of this notebook? {.unlisted .unnumbered}

- Here, I model NPS dot products as a function of cue, stimulus intensity and expectation ratings.
- One of the findings is that low cues lead to higher NPS dotproducts in the high intensity group, and that this effect becomes non-significant across sessions.
- 03/23/2023: For now, I'm grabbing participants that have complete data, i.e. 18 runs, all three sessions.

```{r include=FALSE}
library(car)
library(psych)
library(reshape)
library(PupillometryR)
library(dplyr)
library(tidyselect)
library(tidyr)
library(stringr)
library(lmerTest)
library(gghalves)
library(plyr)
library(ggpubr)
library(r2mlm)
library(effectsize)
library(devtools)
options(es.use_symbols = TRUE) # get nice symbols when printing! (On Windows, requires R >= 4.2.0)
library(EMAtools)
library(emmeans)
library(sjPlot)
library(sjmisc)
library(sjlabelled)
library(DT)
library(raincloudplots)
devtools::source_url("https://gist.githubusercontent.com/benmarwick/2a1bb0133ff568cbe28d/raw/fb53bd97121f7f9ce947837ef1a4c65a73bffb3f/geom_flat_violin.R")
main_dir = dirname(dirname(getwd()))
file.sources = list.files(file.path(main_dir, "scripts/step02_R/utils"),
                          pattern="*.R",
                          full.names=TRUE,
                          ignore.case=TRUE)
sapply(file.sources,source,.GlobalEnv)
```

## common parameters

```{r}
main_dir = dirname(dirname(getwd()))
print(main_dir)
exclude_list <- "sub-0001|sub-0002|sub-0003|sub-0004|sub-0005|sub-0007|sub-0008|sub-0013|sub-0016|sub-0017|sub-0019|sub-0020|sub-0021|sub-0025|sub-0075"
```

```{r}
## load brain extracted values {.unlisted .unnumbered}
extraction_fname = file.path(main_dir,'analysis/fmri/nilearn/signature_canlabcore/roi-painpathway_sub-all_runtype-pvc_event-stimulus.csv')
brain.df <- load_extraction(fname = extraction_fname)
head(brain.df)

## load behavioral data {.unlisted .unnumbered}
beh.df <- df_load_pvc_beh(datadir = file.path(main_dir, 'data', 'beh', 'beh02_preproc'),
             subject_varkey = "src_subject_id",
             iv = "param_stimulus_type",
             dv = "event04_actual_angle",
             exclude = exclude_list)
```

## merge two dataframes {.unlisted .unnumbered}

```{r}
beh.df.sub <- beh.df[,c("sub", "ses", "run", "runtype", "trial", "cuetype", "stimintensity","event02_expect_angle", "event04_actual_angle")]
pvc <- merge(brain.df, beh.df.sub,
                  by.x = c("sub", "ses", "run", "runtype", "trial", "cuetype", "stimintensity"),
                  by.y = c("sub", "ses", "run", "runtype", "trial", "cuetype", "stimintensity")
                  )
pvc.df <- simple_contrast_singletrial(pvc)
```

## plot task x intensity

```{r echo=FALSE}

taskname = "all"
plot_keyword = "stimulusintensity"
ggtitle_phrase =  "(3 tasks x 3 stimulus intensity)"
pvc$task = factor(pvc.df$task)
analysis_dir <-
  file.path(main_dir,"analysis","mixedeffect","model41_painpathway",
            as.character(Sys.Date())) # nolint
dir.create(analysis_dir,showWarnings = FALSE,recursive = TRUE)

for (extract_key in c("Thal_VPLM_R","Thal_VPLM_L", "Thal_MD", "dpIns_L", "dpIns_R", "aMCC_MPFC")) {

  PLOT_KEYS <- list(
    iv1 = "task", iv2 = "stim_ordered", dv = extract_key,
    sub = "sub", sub_mean = "mean_per_sub",  group_mean = "mean_per_sub_norm_mean",
    dv_keyword = extract_key, legend_keyword = "stimulus intensity",  error = "se",
    ggtitle = paste0(
      str_to_title(extract_key), " dot product: ", str_to_title(taskname), ' ', ggtitle_phrase, " (N = ", length(unique(pvc$sub)), ")"
    ),
    legend_title = paste0(str_to_title(extract_key), " - ", str_to_title(plot_keyword)),
    xlab = "",  ylab = paste(extract_key, " (beta value)"),
    ylim = c(-250, 500), task_name = taskname, w = 10, h = 6,
    save_fname = file.path(analysis_dir, 'TEMP.png'),
    color =  c("#4575B4", "#FFA500", "#D73027")
  )
  p <- plot_signature_tworaincloud(signature_key = extract_key, plot_keys = PLOT_KEYS, df = pvc.df)
  print(p)
}
```

## plot cue x intensity

```{r echo=FALSE}

taskname = "all"
plot_keyword = "stimulusintensity"
ggtitle_phrase =  "(2 cue x 3 stimulus intensity)"
pvc$task = factor(pvc.df$task)

dv_keyword = 'NPSpos'
for (extract_key in c("Thal_VPLM_R","Thal_VPLM_L", "Thal_MD", "dpIns_L", "dpIns_R", "aMCC_MPFC")) {

PLOT_KEYS <- list(
    iv1 = "stim_ordered", iv2 = "cue_ordered", dv = extract_key,
    sub = "sub", sub_mean = "mean_per_sub",  group_mean = "mean_per_sub_norm_mean",
    dv_keyword = extract_key, legend_keyword = "cue",  error = "se",
    ggtitle = paste0(
      str_to_title(extract_key), " dot product: ", str_to_title(taskname), ' ', ggtitle_phrase, " (N = ", length(unique(pvc$sub)), ")"
    ),
    legend_title = paste0(str_to_title(extract_key), " - ", str_to_title(plot_keyword)),
    xlab = "Stimulus intensity",  ylab = paste(extract_key, " (beta value)"),
    ylim = c(-250, 500), task_name = taskname, w = 10, h = 6,
    save_fname = file.path(analysis_dir, 'TEMP.png'),
    color =  c("#4274AD", "#C5263A")
  )
  # [ MODEL ] _________________________________________________ # nolint
model_savefname <- file.path(
  analysis_dir,
  paste(
    "lmer_task-",
    taskname,
    "_rating-",
    extract_key,
    "_",
    as.character(Sys.Date()),
    "_cooksd.txt",
    sep = ""
  )
)

cooksd <- lmer_twofactor_cooksd(
  pvc.df,
  taskname,
  "cue_ordered",
  "stim_ordered",
  "stim_ordered",
  dv=extract_key,
  "sub",
  extract_key,
  model_savefname,
  'random_intercept',
  print_lmer_output = FALSE
)
  p <- plot_signature_tworaincloud(signature_key = extract_key, plot_keys = PLOT_KEYS, df = pvc.df[pvc.df$task == "pain",])
  print(p)

  g <- plot_signature_twolineplot(signature_key = extract_key, plot_keys = PLOT_KEYS, df = pvc.df[pvc.df$task == "pain",])
  print(g)

}
```
