==> scripts/step02_R_bookdown/51_iv-6cond_dv-firglasserSPM_ttl2_rois.Rmd <==
# fMRI :: FIR ~ task {#fir_ttl2}

TODO

- load tsv
- concatenate
- per time column, calculate mean and variance
- plot

```{r libraries, message=FALSE, warning=FALSE, include=FALSE, paged.print=FALSE}
library(car)
library(psych)
library(reshape)
library(plyr)
library(dplyr)
library(tidyselect)
library(tidyr)
library(stringr)
library(lmerTest)
library(gghalves)
library(ggpubr)
source("https://gist.githubusercontent.com/benmarwick/2a1bb0133ff568cbe28d/raw/fb53bd97121f7f9ce947837ef1a4c65a73bffb3f/geom_flat_violin.R")
main_dir = dirname(dirname(getwd()))
file.sources = list.files(file.path(main_dir, "scripts/step02_R/utils"),
                          pattern="*.R",
                          full.names=TRUE,
                          ignore.case=TRUE)
sapply(file.sources,source,.GlobalEnv)
```

```{r}

plot_timeseries_onefactor <-  function(df, iv1,  mean, error, xlab, ylab, ggtitle, color) {

n_points <- 100  # Number of points for interpolation
    g <- ggplot(
      data = df,
      aes(
        x = .data[[iv1]],
        y = .data[[mean]],
        group = 1,
        color = color
      ),
      cex.lab = 1.5,
      cex.axis = 2,
      cex.main = 1.5,
      cex.sub = 1.5
    ) +

      geom_errorbar(aes(
        ymin = (.data[[mean]] - .data[[error]]),
        ymax = (.data[[mean]] + .data[[error]]),
        color = color
      ), width = .1, alpha=0.8) +

      geom_line() +
      geom_point(color=color) +
      ggtitle(ggtitle) +
      xlab(xlab) +
      ylab(ylab) +

      theme_classic() +

      theme(aspect.ratio = .6) +
      expand_limits(x = 3.25) +

      scale_color_manual("",
                         values =  color) +
            # scale_fill_manual("",
                         # values =  color) +
      theme(
        legend.position = c(.99, .99),
        legend.justification = c("right", "top"),
        legend.box.just = "right",
        legend.margin = margin(6, 6, 6, 6)
      ) +
      theme(legend.key = element_rect(fill = "white", colour = "white")) +
      theme_bw()

    return(g)
  }
```

### parameters {TODO: ignore}

```{r}

# parameters
main_dir <- dirname(dirname(getwd()))

datadir <- file.path(main_dir, 'analysis/fmri/nilearn/glm/fir')
analysis_folder  = paste0("model52_iv-6cond_dv-firglasserSPM_ttl2")
analysis_dir <-
  file.path(main_dir,
            "analysis",
            "mixedeffect",
            analysis_folder,
            as.character(Sys.Date())) # nolint
dir.create(analysis_dir,
           showWarnings = FALSE,
           recursive = TRUE)
save_dir <- analysis_dir
```

## taskwise stim effect

```{r}
roi_list <- c('rINS', 'TPJ', 'dACC', 'PHG', 'V1', 'SM', 'MT', 'RSC', 'LOC', 'FFC', 'PIT', 'pSTS', 'AIP', 'premotor')
run_types <- c("pain", "vicarious", "cognitive")
  plot_list <- list()
  TR_length <- 42
for (ROI in roi_list) {

    datadir = file.path(main_dir, "analysis/fmri/spm/fir/ttl2par")
 taskname = 'pain'
exclude <- "sub-0001"
filename <- paste0("sub-*",  "*roi-", ROI, "_tr-42.csv")
  common_path <- Sys.glob(file.path(datadir, "sub-*",  filename
  ))
  filter_path <- common_path[!str_detect(common_path, pattern = exclude)]

df <- do.call("rbind.fill", lapply(filter_path, FUN = function(files) {
    read.table(files, header = TRUE, sep = ",")
    }))

for (run_type in run_types) {
  print(run_type)
  filtered_df <- df[!(df$condition == "rating" | df$condition == "cue" | df$runtype != run_type), ]

  parsed_df <- filtered_df %>%
    separate(condition, into = c("cue", "stim"), sep = "_", remove = FALSE)
  # --------------------- subset regions based on ROI ----------------------------
  df_long <- pivot_longer(parsed_df, cols = starts_with("tr"), names_to = "tr_num", values_to = "tr_value")

  # ----------------------------- clean factor -----------------------------------
  df_long$tr_ordered <- factor(
          df_long$tr_num,
          levels = c(paste0("tr", 1:TR_length))
      )
  df_long$stim_ordered <- factor(
          df_long$stim,
          levels = c("stimH", "stimM", "stimL")
      )

  # --------------------------- summary statistics -------------------------------
  subjectwise <- meanSummary(df_long,
                                        c("sub","tr_ordered", "stim_ordered"), "tr_value")
  groupwise <- summarySEwithin(
    data = subjectwise,
    measurevar = "mean_per_sub",
    withinvars = c( "stim_ordered", "tr_ordered"),
    idvar = "sub"
  )
  groupwise$task <- run_type
  # https://stackoverflow.com/questions/29402528/append-data-frames-together-in-a-for-loop/29419402

  # ... Rest of your data processing code ...

  # subset <- groupwise[groupwise$runtype == run_type, ]
  LINEIV1 = "tr_ordered"
  LINEIV2 = "stim_ordered"
  MEAN = "mean_per_sub_norm_mean"
  ERROR = "se"
  dv_keyword = "actual"
  sorted_indices <- order(groupwise$tr_ordered)
  groupwise_sorted <- groupwise[sorted_indices, ]
  p1 <- plot_timeseries_bar(groupwise_sorted,
                            LINEIV1, LINEIV2, MEAN, ERROR,
                            xlab = "TRs",
                            ylab = paste0(ROI, " activation (A.U.)"),
                            ggtitle = paste0(ROI, ": ",run_type, " (N = ", length(unique(subjectwise$sub)),") time series, Epoch - stimulus"),
                            color = c("#5f0f40","#ae2012", "#fcbf49"))
  time_points <- seq(1, 0.46 * TR_length, 0.46)
  #p1 <- p1 + scale_x_discrete(labels = setNames(time_points, colnames(df_long)[7:(7 + TR_length)])) + theme_classic()

  plot_list[[run_type]] <- p1 + theme_classic()
}

  # --------------------------- plot three tasks -------------------------------
library(gridExtra)
plot_list <- lapply(plot_list, function(plot) {
  plot + theme(plot.margin = margin(5, 5, 5, 5))  # Adjust plot margins if needed
})
combined_plot <- ggpubr::ggarrange(plot_list[["pain"]],plot_list[["vicarious"]],plot_list[["cognitive"]],
                  common.legend = TRUE,legend = "bottom", ncol = 3, nrow = 1,
                  widths = c(3, 3, 3), heights = c(.5,.5,.5), align = "v")
combined_plot
ggsave(file.path(save_dir, paste0("roi-", ROI,"_epoch-stim_desc-highstimGTlowstim.png")), combined_plot, width = 12, height = 4)

}

```

#### PCA subjectwise

```{r}
# install.packages("ggplot2")    # Install ggplot2 if you haven't already
# install.packages("FactoMineR") # Install FactoMineR if you haven't already
library(ggplot2)
library(FactoMineR)
run_types = c("pain")
for (run_type in run_types) {
  print(run_type)
  filtered_df <- df[!(df$condition == "rating" | df$condition == "cue" | df$runtype != run_type), ]

  parsed_df <- filtered_df %>%
    separate(condition, into = c("cue", "stim"), sep = "_", remove = FALSE)
  # --------------------- subset regions based on ROI ----------------------------
  df_long <- pivot_longer(parsed_df, cols = starts_with("tr"), names_to = "tr_num", values_to = "tr_value")

  # ----------------------------- clean factor -----------------------------------
  df_long$tr_ordered <- factor(
          df_long$tr_num,
          levels = c(paste0("tr", 1:TR_length))
      )
  df_long$stim_ordered <- factor(
          df_long$stim,
          levels = c("stimH", "stimM", "stimL")
      )

  # --------------------------- summary statistics -------------------------------
  subjectwise <- meanSummary(df_long,
                                        c("sub","tr_ordered", "stim_ordered"), "tr_value")

# Assuming your original dataframe is named 'df'

# Convert the dataframe to wide format
df_wide <- pivot_wider(subjectwise,
                       id_cols = c("tr_ordered", "stim_ordered"),
                       names_from = c("sub"),
                       values_from = "mean_per_sub")

# df_wide <- pivot_wider(subjectwise,
#                        id_cols = c("sub", "ROIindex","stim_ordered"),
#                        names_from = "tr_ordered",
#                        values_from = "mean_per_sub")
stim_high.df <- df_wide[df_wide$stim_ordered == "stimH",]
stim_med.df <- df_wide[df_wide$stim_ordered == "stimM",]
stim_low.df <- df_wide[df_wide$stim_ordered == "stimL",]
# selected_columns <- subset(stim_high.df, select = 2:(ncol(stim_high.df) - 1))
meanhighdf <- data.frame(subset(stim_high.df, select = 3:(ncol(stim_high.df) - 1)))
high.pca_result <- prcomp(meanhighdf)
high.pca_scores <- as.data.frame(high.pca_result$x)
# Access the proportion of variance explained by each principal component
high.variance_explained <- high.pca_result$sdev^2 / sum(high.pca_result$sdev^2)
plot(high.variance_explained)
# Access the standard deviations of each principal component
high.stdev <- high.pca_result$sdev

meanmeddf <- data.frame(subset(stim_med.df, select = 3:(ncol(stim_med.df) - 1)))
med.pca <- prcomp(meanmeddf)
med.pca_scores <- as.data.frame(med.pca$x)

meanlowdf <- data.frame(subset(stim_low.df, select = 3:(ncol(stim_low.df) - 1)))
low.pca <- prcomp(meanlowdf)
low.pca_scores <- as.data.frame(low.pca$x)
library(plotly)  # You can use plotly to create an interactive 3D plot
# plot_ly(high.pca_scores, x = ~PC1, y = ~PC2, z = ~PC3, type = "scatter3d", mode = "markers")
# plot_ly(low.pca_scores, x = ~PC1, y = ~PC2, z = ~PC3, type = "scatter3d", mode = "markers")
combined_pca_scores <- rbind(high.pca_scores, med.pca_scores, low.pca_scores)

# Add a new column to indicate the stim_ordered category (high_stim or low_stim)
combined_pca_scores$stim_ordered <- c(rep("high_stim", nrow(high.pca_scores)),
                                      rep("med_stim", nrow(med.pca_scores)),
                                      rep("low_stim", nrow(low.pca_scores)))

# Create the 3D PCA plot
plot_ly(combined_pca_scores, x = ~PC1, y = ~PC2, z = ~PC3, type = "scatter3d", mode = "markers",
        color = ~stim_ordered)
# data_matrix <- groupwise[groupwise$stim_ordered == "high_stim",c("tr_ordered", "mean_per_sub_norm_mean")]
# sorted_indices <- order(data_matrix$tr_ordered)
# df_ordered <- data_matrix[sorted_indices, ]
# pca_result <- PCA(data_matrix$mean_per_sub_norm_mean)
# datapoints <- df$datapoints
}
```

```{r}
plot_ly(combined_pca_scores, x = ~PC1, y = ~PC2, z = ~PC3, type = "scatter3d", mode = "markers",
        color = ~stim_ordered)
```

#### PCA groupwise

```{r}
# install.packages("ggplot2")    # Install ggplot2 if you haven't already
# install.packages("FactoMineR") # Install FactoMineR if you haven't already
library(ggplot2)
library(FactoMineR)



# Assuming your original dataframe is named 'df'

# Convert the dataframe to wide format
df_wide.group <- pivot_wider(subjectwise,
                       id_cols = c("tr_ordered", "stim_ordered"),
                       names_from = "sub",
                       values_from = "mean_per_sub")
# ------
# data_matrix <- groupwise[groupwise$stim_ordered == "high_stim",c("tr_ordered", "mean_per_sub_norm_mean")]
# sorted_indices <- order(data_matrix$tr_ordered)
# df_ordered <- data_matrix[sorted_indices, ]
# datapoints <- df_ordered$mean_per_sub_norm_mean
# data_df <- data.frame(Dim1 = datapoints, Dim2 = datapoints, Dim3 = datapoints)
# pca <- prcomp(data_df)
# pca_scores <- as.data.frame(pca$x)
# plot_ly(pca_scores, x = ~PC1, y = ~PC2, z = ~PC3, type = "scatter3d", mode = "markers")
# -------
stim_high.df <- df_wide[df_wide$stim_ordered == "stimH",]
stim_low.df <- df_wide[df_wide$stim_ordered == "stimL",]
# selected_columns <- subset(stim_high.df, select = 2:(ncol(stim_high.df) - 1))
meanhighdf <- data.frame(subset(stim_high.df, select = 3:(ncol(stim_high.df) - 1)))
high.pca <- prcomp(meanhighdf)
high.pca_scores <- as.data.frame(high.pca$x)

meanlowdf <- data.frame(subset(stim_low.df, select = 3:(ncol(stim_low.df) - 1)))
low.pca <- prcomp(meanlowdf)
low.pca_scores <- as.data.frame(low.pca$x)
library(plotly)  # You can use plotly to create an interactive 3D plot
# plot_ly(high.pca_scores, x = ~PC1, y = ~PC2, z = ~PC3, type = "scatter3d", mode = "markers")
# plot_ly(low.pca_scores, x = ~PC1, y = ~PC2, z = ~PC3, type = "scatter3d", mode = "markers")
combined_pca_scores <- rbind(high.pca_scores, low.pca_scores)

# Add a new column to indicate the stim_ordered category (high_stim or low_stim)
combined_pca_scores$stim_ordered <- c(rep("high_stim", nrow(high.pca_scores)), rep("low_stim", nrow(low.pca_scores)))

# Create the 3D PCA plot
plot_ly(combined_pca_scores, x = ~PC1, y = ~PC2, z = ~PC3, type = "scatter3d", mode = "markers",
        color = ~stim_ordered)
# data_matrix <- groupwise[groupwise$stim_ordered == "high_stim",c("tr_ordered", "mean_per_sub_norm_mean")]
# sorted_indices <- order(data_matrix$tr_ordered)
# df_ordered <- data_matrix[sorted_indices, ]
# pca_result <- PCA(data_matrix$mean_per_sub_norm_mean)
# datapoints <- df$datapoints

# Assuming you have a dataframe named 'data' containing the 20 data points, 'x' and 'y' values, and corresponding standard deviations 'sd'

# Load the ggplot2 library
# install.packages("ggplot2")
library(ggplot2)

# Create the plot
# y = "mean_per_sub_mean"z
# combined_pca <- combined_pca_scores %>%
  # mutate(group_index = group_indices(., stim_ordered))

combined_pca <- combined_pca_scores %>%
  group_by(stim_ordered) %>%
  mutate(group_index = row_number())
ggplot(combined_pca, aes(x=group_index,y=PC1, group = stim_ordered, colour=stim_ordered)) +
  stat_smooth(method="loess", span=0.25, se=TRUE, aes(color=stim_ordered), alpha=0.3) +
  theme_bw()


```

```{r}
# Assuming you have a dataframe named 'data' containing the 20 data points, 'x' and 'y' values, and corresponding standard deviations 'sd'

# Load the ggplot2 library
# install.packages("ggplot2")
library(ggplot2)

# Create the plot
# y = "mean_per_sub_mean"z
ggplot(groupwise, aes(x=tr_ordered,y=mean_per_sub_mean, group = stim_ordered, colour=stim_ordered)) +
  stat_smooth(method="loess", span=0.25, se=TRUE, aes(color=stim_ordered), alpha=0.3) +
  theme_bw()

```

```{r}
# ggplot(data=groupwise, aes(x=tr_ordered, y=mean_per_sub_mean, ymin=se, ymax=se, fill=stim_ordered, linetype=stim_ordered)) +
#  geom_line() +
#  geom_ribbon(alpha=0.5)
# Assuming you have a dataframe named 'data' containing the 20 mean data points and corresponding standard errors
# 'x' represents the x-values (e.g., time points)
# 'mean_y' represents the mean y-values
# 'se_y' represents the standard errors of the mean y-values

# Load the ggplot2 library
# install.packages("ggplot2")
library(ggplot2)
# groupwise$x <- as.numeric(groupwise$x)
#
# # Sort the dataframe by the 'x' variable (if it's not already sorted)
# data <- data[order(data$x), ]

# Create the plot
# Create the plot with custom span and smoothing method
ggplot(groupwise, aes(x=tr_ordered,y=mean_per_sub_mean)) +
  geom_line() +                                   # Plot the smooth line for the mean
  geom_ribbon(aes(ymin = mean_per_sub_mean - se, ymax = mean_per_sub_mean + se), alpha = 0.3) + # Add the ribbon for standard error
  geom_smooth(method = "loess", span = 0.1, se = FALSE) +       # Add the loess smoothing curve
  labs(x = "X-axis Label", y = "Y-axis Label", title = "Smooth Line with Standard Error Ribbon") +
  theme_minimal()

```

## DEP: epoch: stim, high cue vs low cue

```{r}
# filtered_df <- subset(df, condition != "rating")
filtered_df <- df[!(df$condition == "rating" | df$condition == "cue"), ]

parsed_df <- filtered_df %>%
  separate(condition, into = c("cue", "stim"), sep = "_", remove = FALSE)

TR_length <- 42
# --------------------- subset regions based on ROI ----------------------------
df_long <- pivot_longer(parsed_df, cols = starts_with("tr"), names_to = "tr_num", values_to = "tr_value")

# ----------------------------- clean factor -----------------------------------
df_long$tr_ordered <- factor(
        df_long$tr_num,
        levels = c(paste0("tr", 1:TR_length))
    )
df_long$cue_ordered <- factor(
        df_long$cue,
        levels = c("cueH","cueL")
    )

# --------------------------- summary statistics -------------------------------
subjectwise <- meanSummary(df_long,
                                      c("sub", "tr_ordered", "cue_ordered"), "tr_value")
groupwise <- summarySEwithin(
  data = subjectwise,
  measurevar = "mean_per_sub",
  withinvars = c("cue_ordered", "tr_ordered"),
  idvar = "sub"
)
groupwise$task <- taskname
# https://stackoverflow.com/questions/29402528/append-data-frames-together-in-a-for-loop/29419402

# --------------------------------- plot ---------------------------------------
LINEIV1 = "tr_ordered"
LINEIV2 = "cue_ordered"
MEAN = "mean_per_sub_norm_mean"
ERROR = "se"
dv_keyword = "actual"
sorted_indices <- order(groupwise$tr_ordered)
groupwise_sorted <- groupwise[sorted_indices, ]
p1 = plot_timeseries_bar(groupwise,
               LINEIV1, LINEIV2, MEAN, ERROR,  xlab = "Runs" , ylab= "Epoch: stimulus, High cue vs. Low cue", ggtitle="time_series", color=c("red", "blue"))
time_points <- seq(1, 0.46 * TR_length, 0.46)
p1 + scale_x_discrete(labels = setNames(time_points, colnames(df_long)[7:7+TR_length]))+ theme_classic()

```

## taskwise cue effect

```{r}
roi_list <- c('rINS', 'TPJ', 'dACC', 'PHG', 'V1', 'SM', 'MT', 'RSC', 'LOC', 'FFC', 'PIT', 'pSTS', 'AIP', 'premotor')
for (ROI in roi_list) {

    datadir = file.path(main_dir, "analysis/fmri/spm/fir/ttl2par")
# taskname = 'pain'
exclude <- "sub-0001"
filename <- paste0("sub-*",  "*roi-", ROI, "_tr-42.csv")
  common_path <- Sys.glob(file.path(datadir, "sub-*",  filename
  ))
  filter_path <- common_path[!str_detect(common_path, pattern = exclude)]

df <- do.call("rbind.fill", lapply(filter_path, FUN = function(files) {
    read.table(files, header = TRUE, sep = ",")
    }))


run_types <- c("pain", "vicarious", "cognitive")
  plot_list <- list()
  TR_length <- 42
for (run_type in run_types) {
  filtered_df <- df[!(df$condition == "rating" | df$condition == "cue" | df$runtype != run_type), ]

  parsed_df <- filtered_df %>%
    separate(condition, into = c("cue", "stim"), sep = "_", remove = FALSE)
  # --------------------- subset regions based on ROI ----------------------------
  df_long <- pivot_longer(parsed_df, cols = starts_with("tr"), names_to = "tr_num", values_to = "tr_value")

  # ----------------------------- clean factor -----------------------------------
  df_long$tr_ordered <- factor(
          df_long$tr_num,
          levels = c(paste0("tr", 1:TR_length))
      )
df_long$cue_ordered <- factor(
        df_long$cue,
        levels = c("cueH","cueL")
    )

  # --------------------------- summary statistics -------------------------------
  subjectwise <- meanSummary(df_long,
                                        c("sub","tr_ordered", "cue_ordered"), "tr_value")
  groupwise <- summarySEwithin(
    data = subjectwise,
    measurevar = "mean_per_sub",
    withinvars = c( "cue_ordered", "tr_ordered"),
    idvar = "sub"
  )
  groupwise$task <- run_type
  # https://stackoverflow.com/questions/29402528/append-data-frames-together-in-a-for-loop/29419402

  # ... Rest of your data processing code ...

  # subset <- groupwise[groupwise$runtype == run_type, ]
  LINEIV1 = "tr_ordered"
  LINEIV2 = "cue_ordered"
  MEAN = "mean_per_sub_norm_mean"
  ERROR = "se"
  dv_keyword = "actual"
  sorted_indices <- order(groupwise$tr_ordered)
  groupwise_sorted <- groupwise[sorted_indices, ]
  p1 <- plot_timeseries_bar(groupwise_sorted,
                            LINEIV1, LINEIV2, MEAN, ERROR,
                            xlab = "TRs",
                            ylab = paste0(ROI, " activation (A.U.)"),
                            ggtitle = paste0(run_type, " time series, Epoch - stimulus"),
                            color =c("red", "blue"))
  time_points <- seq(1, 0.46 * TR_length, 0.46)
  #p1 <- p1 + scale_x_discrete(labels = setNames(time_points, colnames(df_long)[7:(7 + TR_length)])) + theme_classic()

  plot_list[[run_type]] <- p1 + theme_classic()
}

  # --------------------------- plot three tasks -------------------------------
library(gridExtra)
plot_list <- lapply(plot_list, function(plot) {
  plot + theme(plot.margin = margin(5, 5, 5, 5))  # Adjust plot margins if needed
})
combined_plot <- ggpubr::ggarrange(plot_list[["pain"]],plot_list[["vicarious"]],plot_list[["cognitive"]],
                  common.legend = TRUE,legend = "bottom", ncol = 3, nrow = 1,
                  widths = c(3, 3, 3), heights = c(.5,.5,.5), align = "v")
combined_plot
ggsave(file.path(save_dir, paste0("roi-", ROI, "_epoch-stim_desc-highcueGTlowcue.png")), combined_plot, width = 12, height = 4)
}
```

## epoch: stim, rating

```{r eval=FALSE, include=FALSE}
# ------------------------------------------------------------------------------
#                       epoch stim, high cue vs low cue
# ------------------------------------------------------------------------------
# --------------------- subset regions based on ROI ----------------------------
taskname = 'pain'

parsed_df <- df[(df$condition == "rating"), ]
TR_length <- 42
df_rating <- pivot_longer(parsed_df, cols = starts_with("tr"), names_to = "tr_num", values_to = "tr_value")

# ----------------------------- clean factor -----------------------------------
df_rating$tr_ordered <- factor(
        df_rating$tr_num,
        levels = c(paste0("tr", 1:TR_length))
    )

# --------------------------- summary statistics -------------------------------
subjectwise <- meanSummary(df_rating,
                                      c("sub", "tr_ordered"), "tr_value")
groupwise <- summarySEwithin(
  data = subjectwise,
  measurevar = "mean_per_sub",
  withinvars = c( "tr_ordered"),
  idvar = "sub"
)
groupwise$task <- taskname
# https://stackoverflow.com/questions/29402528/append-data-frames-together-in-a-for-loop/29419402



# Assuming your data frame is named "time_series_data"

# Create the ggplot
gg <- ggplot(groupwise, aes(x = tr_ordered, y = mean_per_sub_norm_mean, group = 1)) +
  geom_line() +
  geom_point() +
  geom_errorbar(aes(ymin = mean_per_sub_norm_mean - se, ymax = mean_per_sub_norm_mean + se), width = 0.2) +
  labs(x = "Time", y = "Amplitude", title = "Time Series Data with Error Bars")
gg <- gg + theme_classic() +       theme(legend.key = element_rect(fill = "white", colour = "white")) +
      theme_bw()
# Print the ggplot
print(gg)

# --------------------------------- plot ---------------------------------------
LINEIV1 = "tr_ordered"
# LINEIV2 = "cue_ordered"
MEAN = "mean_per_sub_norm_mean"
ERROR = "se"
dv_keyword = "actual"
sorted_indices <- order(groupwise$tr_ordered)
groupwise_sorted <- groupwise[sorted_indices, ]
p2 = plot_timeseries_onefactor(groupwise_sorted,
               LINEIV1,MEAN, ERROR,  xlab = "Runs" , ylab= "Epoch: stimulus, High cue vs. Low cue", ggtitle="time_series", color="red")
time_points <- seq(1, 0.46 * TR_length, 0.46)
p2 + scale_x_discrete(labels = setNames(time_points, colnames(df_long)[7:7+TR_length]))+ theme_classic()
p2 + theme_classic()
```

## epoch: 6 cond

```{r}
# ------------------------------------------------------------------------------
#                       epoch stim, high cue vs low cue
# ------------------------------------------------------------------------------
# --------------------- subset regions based on ROI ----------------------------

# ----------------------------- clean factor -----------------------------------
df_long$tr_ordered <- factor(
        df_long$tr_num,
        levels = c(paste0("tr", 1:TR_length))
    )
df_long$cue_ordered <- factor(
        df_long$cue,
        levels = c("cueH", "cueL")
    )
df_long$stim_ordered <- factor(
        df_long$stim,
        levels = c("stimH", "stimM", "stimL")
    )

df_long$sixcond <- factor(
        df_long$condition,
        levels = c("cueH_stimH", "cueL_stimH",
                   "cueH_stimM", "cueL_stimM",
                   "cueH_stimL", "cueL_stimL")
)
# --------------------------- summary statistics -------------------------------
subjectwise <- meanSummary(df_long,
                                      c("sub", "tr_ordered", "sixcond"), "tr_value")
groupwise <- summarySEwithin(
  data = subjectwise,
  measurevar = "mean_per_sub",
  withinvars = c("sixcond", "tr_ordered"),
  idvar = "sub"
)
groupwise$task <- taskname
# https://stackoverflow.com/questions/29402528/append-data-frames-together-in-a-for-loop/29419402

# --------------------------------- plot ---------------------------------------
LINEIV1 = "tr_ordered"
LINEIV2 = "sixcond"
MEAN = "mean_per_sub_norm_mean"
ERROR = "se"
dv_keyword = "actual"
sorted_indices <- order(groupwise$tr_ordered)
groupwise_sorted <- groupwise[sorted_indices, ]
p3H = plot_timeseries_bar(groupwise,
               LINEIV1, LINEIV2, MEAN, ERROR,  xlab = "Runs" , ylab= "Epoch: stimulus, High cue vs. Low cue", ggtitle=paste0("High intensity - Low cue vs. High cue (N = ", unique(groupwise$N), ")" ), color=c("red","#5f0f40","gray", "gray", "gray", "gray"))
time_points <- seq(1, 0.46 * TR_length, 0.46)
p3H + scale_x_discrete(labels = setNames(time_points, colnames(df_long)[7:7+TR_length]))+ theme_classic()
p3H + theme_classic()
```

```{r}
p3M = plot_timeseries_bar(groupwise,
               LINEIV1, LINEIV2, MEAN, ERROR,  xlab = "Runs" , ylab= "Epoch: stimulus, High cue vs. Low cue", ggtitle=paste0("Medium intensity - Low cue vs. High cue (N = ", unique(groupwise$N), ")"), color=c("#d6d6d6","#d6d6d6","#bc3908", "#f6aa1c", "gray", "gray"))
time_points <- seq(1, 0.46 * TR_length, 0.46)
p3M + scale_x_discrete(labels = setNames(time_points, colnames(df_long)[7:7+TR_length]))+ theme_classic()
p3M + theme_classic()
```

```{r}
p3L = plot_timeseries_bar(groupwise,
               LINEIV1, LINEIV2, MEAN, ERROR,  xlab = "Runs" , ylab= "Epoch: stimulus, High cue vs. Low cue", ggtitle=paste0("Low intensity - Low cue vs. High cue (N = ", unique(groupwise$N), ")"), color=c("gray","gray","gray", "gray", "#2541b2", "#00a6fb"))
time_points <- seq(1, 0.46 * TR_length, 0.46)
p3L + scale_x_discrete(labels = setNames(time_points, colnames(df_long)[7:7+TR_length]))+ theme_classic()
p3L + theme_classic()
```

## taskwise 6 cond effect

```{r}
# ------------------------------------------------------------------------------
#                       epoch stim, high cue vs low cue
# ------------------------------------------------------------------------------
# --------------------- subset regions based on ROI ----------------------------
run_types <- c("pain", "vicarious", "cognitive")

  TR_length <- 42
for (run_type in run_types) {
  filtered_df <- df[!(df$condition == "rating" | df$condition == "cue" | df$runtype != run_type), ]
plot_list <- list()

  parsed_df <- filtered_df %>%
    separate(condition, into = c("cue", "stim"), sep = "_", remove = FALSE)
  # --------------------- subset regions based on ROI ----------------------------
  df_long <- pivot_longer(parsed_df, cols = starts_with("tr"), names_to = "tr_num", values_to = "tr_value")

# ----------------------------- clean factor -----------------------------------
df_long$tr_ordered <- factor(
        df_long$tr_num,
        levels = c(paste0("tr", 1:TR_length))
    )
df_long$cue_ordered <- factor(
        df_long$cue,
        levels = c("cueH", "cueL")
    )
df_long$stim_ordered <- factor(
        df_long$stim,
        levels = c("stimH", "stimM", "stimL")
    )

df_long$sixcond <- factor(
        df_long$condition,
        levels = c("cueH_stimH", "cueL_stimH",
                   "cueH_stimM", "cueL_stimM",
                   "cueH_stimL", "cueL_stimL")
)
# --------------------------- summary statistics -------------------------------
subjectwise <- meanSummary(df_long,
                                      c("sub", "tr_ordered", "sixcond"), "tr_value")
groupwise <- summarySEwithin(
  data = subjectwise,
  measurevar = "mean_per_sub",
  withinvars = c("sixcond", "tr_ordered"),
  idvar = "sub"
)
groupwise$task <- taskname
# https://stackoverflow.com/questions/29402528/append-data-frames-together-in-a-for-loop/29419402

# --------------------------------- plot ---------------------------------------
LINEIV1 = "tr_ordered"
LINEIV2 = "sixcond"
MEAN = "mean_per_sub_norm_mean"
ERROR = "se"
dv_keyword = "actual"
sorted_indices <- order(groupwise$tr_ordered)
groupwise_sorted <- groupwise[sorted_indices, ]
p3H = plot_timeseries_bar(groupwise,
               LINEIV1, LINEIV2, MEAN, ERROR,  xlab = "Runs" , ylab= "Epoch: stimulus, High cue vs. Low cue", ggtitle=paste0("High intensity - Low cue vs. High cue (N = ", unique(groupwise$N), ")" ), color=c("red","#5f0f40","gray", "gray", "gray", "gray"))
time_points <- seq(1, 0.46 * TR_length, 0.46)
# p3H + scale_x_discrete(labels = setNames(time_points, colnames(df_long)[7:7+TR_length]))+ theme_classic()
p3H + theme_classic()
plot_list[["H"]] <- p3H + theme_classic()

p3M = plot_timeseries_bar(groupwise,
               LINEIV1, LINEIV2, MEAN, ERROR,  xlab = "Runs" , ylab= "Epoch: stimulus, High cue vs. Low cue", ggtitle=paste0("Medium intensity - Low cue vs. High cue (N = ", unique(groupwise$N), ")"), color=c("#d6d6d6","#d6d6d6","#bc3908", "#f6aa1c", "gray", "gray"))
time_points <- seq(1, 0.46 * TR_length, 0.46)
# p3M + scale_x_discrete(labels = setNames(time_points, colnames(df_long)[7:7+TR_length]))+ theme_classic()
plot_list[["M"]] <- p3M + theme_classic()

p3L = plot_timeseries_bar(groupwise,
               LINEIV1, LINEIV2, MEAN, ERROR,  xlab = "Runs" , ylab= "Epoch: stimulus, High cue vs. Low cue", ggtitle=paste0("Low intensity - Low cue vs. High cue (N = ", unique(groupwise$N), ")"), color=c("gray","gray","gray", "gray", "#2541b2", "#00a6fb"))
time_points <- seq(1, 0.46 * TR_length, 0.46)
# p3L + scale_x_discrete(labels = setNames(time_points, colnames(df_long)[7:7+TR_length]))+ theme_classic()
plot_list[["L"]] <- p3L + theme_classic()


  # --------------------------- plot three tasks -------------------------------
library(gridExtra)
plot_list <- lapply(plot_list, function(plot) {
  plot + theme(plot.margin = margin(5, 5, 5, 5))  # Adjust plot margins if needed
})
combined_plot <- ggpubr::ggarrange(plot_list[["H"]],plot_list[["M"]],plot_list[["L"]],
                  common.legend = FALSE,legend = "bottom", ncol = 3, nrow = 1,
                  widths = c(3, 3, 3), heights = c(.5,.5,.5), align = "v")
combined_plot
ggsave(file.path(save_dir, paste0("taskwise-",run_type, "_epoch-stim_desc-stimcuecomparison.png")), combined_plot, width = 12, height = 4)
}
```



```{r}
# ------------------------------------------------------------------------------
#                    3x2 GRID: CUE EPOCH + STIM EPOCH
# ------------------------------------------------------------------------------

run_types <- c("pain", "vicarious", "cognitive")
TR_length <- 42

for (run_type in run_types) {
  
  # ============================================================================
  # PREPARE DATA - Get both cue and stim epochs
  # ============================================================================
  
  # Get STIM epoch data (exclude rating and cue conditions)
  filtered_df_stim <- df[!(df$condition == "rating" | | df$runtype != run_type), ]
  
  # Get CUE epoch data (only cue condition, but we need to match it to trial info)
  # Since cue condition doesn't have stim info, we need to get it from the trial structure
  # Assuming your data has a way to link cue trials to their corresponding stim conditions
  
  # If your dataframe has cue trials as separate rows with condition == "cue",
  # but they're linked to stim trials through trial numbers or similar:
  filtered_df_cue <- df[df$condition == "cue" & df$runtype == run_type, ]
  
  # If cue trials don't exist as separate rows, we'll use the stim trials 
  # but select only the cue epoch TRs (typically the first few TRs)
  # Let me show both approaches:
  
  # ============================================================================
  # APPROACH: Use stim epoch data but separate by TR timing
  # ============================================================================
  # Assuming TRs 1-4 are cue epoch, TRs 5-42 are stim epoch (adjust as needed)
  
  cue_TR_range <- 1:42  # Adjust based on your experimental design
  stim_TR_range <- 1:42  # Adjust based on your experimental design
  
  # Prepare stim data
  parsed_df_stim <- filtered_df_stim %>%
    separate(condition, into = c("cue", "stim"), sep = "_", remove = FALSE)
  
  df_long_stim <- pivot_longer(parsed_df_stim, 
                               cols = starts_with("tr"), 
                               names_to = "tr_num", 
                               values_to = "tr_value")
  
  # Clean factors for STIM
  df_long_stim$tr_ordered <- factor(df_long_stim$tr_num, 
                                     levels = c(paste0("tr", 1:TR_length)))
  df_long_stim$sixcond <- factor(df_long_stim$condition,
                                  levels = c("cueH_stimH", "cueL_stimH",
                                           "cueH_stimM", "cueL_stimM",
                                           "cueH_stimL", "cueL_stimL"))
  
  # Filter for STIM epoch TRs only
  df_long_stim <- df_long_stim %>%
    filter(tr_num %in% paste0("tr", stim_TR_range))
  
  # Prepare CUE data (same df but filter for cue TRs)
  df_long_cue <- pivot_longer(parsed_df_stim, 
                               cols = starts_with("tr"), 
                               names_to = "tr_num", 
                               values_to = "tr_value")
  
  df_long_cue$tr_ordered <- factor(df_long_cue$tr_num, 
                                   levels = c(paste0("tr", 1:TR_length)))
  df_long_cue$sixcond <- factor(df_long_cue$condition,
                                levels = c("cueH_stimH", "cueL_stimH",
                                         "cueH_stimM", "cueL_stimM",
                                         "cueH_stimL", "cueL_stimL"))
  
  # Filter for CUE epoch TRs only
  df_long_cue <- df_long_cue %>%
    filter(tr_num %in% paste0("tr", cue_TR_range))
  
  # ============================================================================
  # SUMMARY STATISTICS - STIM EPOCH
  # ============================================================================
  
  subjectwise_stim <- meanSummary(df_long_stim, 
                                  c("sub", "tr_ordered", "sixcond"), 
                                  "tr_value")
  groupwise_stim <- summarySEwithin(
    data = subjectwise_stim,
    measurevar = "mean_per_sub",
    withinvars = c("sixcond", "tr_ordered"),
    idvar = "sub"
  )
  groupwise_stim$task <- taskname
  
  # ============================================================================
  # SUMMARY STATISTICS - CUE EPOCH
  # ============================================================================
  
  subjectwise_cue <- meanSummary(df_long_cue, 
                                 c("sub", "tr_ordered", "sixcond"), 
                                 "tr_value")
  groupwise_cue <- summarySEwithin(
    data = subjectwise_cue,
    measurevar = "mean_per_sub",
    withinvars = c("sixcond", "tr_ordered"),
    idvar = "sub"
  )
  groupwise_cue$task <- taskname
  
  # ============================================================================
  # PLOTTING PARAMETERS
  # ============================================================================
  
  LINEIV1 = "tr_ordered"
  LINEIV2 = "sixcond"
  MEAN = "mean_per_sub_norm_mean"
  ERROR = "se"
  
  # ============================================================================
  # ROW 1: HIGH INTENSITY
  # ============================================================================
  
  # CUE epoch - High intensity (cueH_stimH vs cueL_stimH)
  p_cue_H <- plot_timeseries_bar(groupwise_cue,
                LINEIV1, LINEIV2, MEAN, ERROR,
                xlab = "TRs (Cue Epoch)", 
                ylab = "Activation (a.u.)",
                ggtitle = paste0("CUE: High Intensity (N = ", unique(groupwise_cue$N), ")"),
                color = c("red", "#5f0f40", "gray", "gray", "gray", "gray")) + 
    theme_classic()
  
  # STIM epoch - High intensity
  p_stim_H <- plot_timeseries_bar(groupwise_stim,
                LINEIV1, LINEIV2, MEAN, ERROR,
                xlab = "TRs (Stim Epoch)", 
                ylab = "Activation (a.u.)",
                ggtitle = paste0("STIM: High Intensity (N = ", unique(groupwise_stim$N), ")"),
                color = c("red", "#5f0f40", "gray", "gray", "gray", "gray")) + 
    theme_classic()
  
  # ============================================================================
  # ROW 2: MEDIUM INTENSITY
  # ============================================================================
  
  # CUE epoch - Medium intensity (cueH_stimM vs cueL_stimM)
  p_cue_M <- plot_timeseries_bar(groupwise_cue,
                LINEIV1, LINEIV2, MEAN, ERROR,
                xlab = "TRs (Cue Epoch)", 
                ylab = "Activation (a.u.)",
                ggtitle = paste0("CUE: Medium Intensity (N = ", unique(groupwise_cue$N), ")"),
                color = c("#d6d6d6", "#d6d6d6", "#bc3908", "#f6aa1c", "gray", "gray")) + 
    theme_classic()
  
  # STIM epoch - Medium intensity
  p_stim_M <- plot_timeseries_bar(groupwise_stim,
                LINEIV1, LINEIV2, MEAN, ERROR,
                xlab = "TRs (Stim Epoch)", 
                ylab = "Activation (a.u.)",
                ggtitle = paste0("STIM: Medium Intensity (N = ", unique(groupwise_stim$N), ")"),
                color = c("#d6d6d6", "#d6d6d6", "#bc3908", "#f6aa1c", "gray", "gray")) + 
    theme_classic()
  
  # ============================================================================
  # ROW 3: LOW INTENSITY
  # ============================================================================
  
  # CUE epoch - Low intensity (cueH_stimL vs cueL_stimL)
  p_cue_L <- plot_timeseries_bar(groupwise_cue,
                LINEIV1, LINEIV2, MEAN, ERROR,
                xlab = "TRs (Cue Epoch)", 
                ylab = "Activation (a.u.)",
                ggtitle = paste0("CUE: Low Intensity (N = ", unique(groupwise_cue$N), ")"),
                color = c("gray", "gray", "gray", "gray", "#2541b2", "#00a6fb")) + 
    theme_classic()
  
  # STIM epoch - Low intensity
  p_stim_L <- plot_timeseries_bar(groupwise_stim,
                LINEIV1, LINEIV2, MEAN, ERROR,
                xlab = "TRs (Stim Epoch)", 
                ylab = "Activation (a.u.)",
                ggtitle = paste0("STIM: Low Intensity (N = ", unique(groupwise_stim$N), ")"),
                color = c("gray", "gray", "gray", "gray", "#2541b2", "#00a6fb")) + 
    theme_classic()
  
  # ============================================================================
  # COMBINE INTO 3x2 GRID
  # ============================================================================
  
  library(gridExtra)
  library(ggpubr)
  
  # List of all plots
  plots <- list(p_cue_H, p_stim_H,    # Row 1
                p_cue_M, p_stim_M,    # Row 2
                p_cue_L, p_stim_L)    # Row 3
  
  # Adjust margins
  plots <- lapply(plots, function(plot) {
    plot + theme(plot.margin = margin(5, 5, 5, 5))
  })
  
  # Arrange in 3 rows x 2 columns
  combined_plot <- ggarrange(
    plots[[1]], plots[[2]],  # Row 1: High intensity
    plots[[3]], plots[[4]],  # Row 2: Medium intensity
    plots[[5]], plots[[6]],  # Row 3: Low intensity
    common.legend = FALSE,
    legend = "bottom",
    ncol = 2, 
    nrow = 3,
    widths = c(1, 1),
    heights = c(1, 1, 1),
    align = "hv",
    labels = c("A. Cue", "B. Stimulus", "", "", "", ""),
    font.label = list(size = 12, face = "bold")
  )
  
  combined_plot
  
  # Save
  ggsave(
    file.path(save_dir, 
              paste0("taskwise-", run_type, 
                    "_epoch-cue-and-stim_3x2grid.png")),
    combined_plot, 
    width = 14, 
    height = 12
  )
}
```


# cue shift?
```{r}
# ------------------------------------------------------------------------------
#                    epoch stim, high cue vs low cue BY RUN ORDER
# ------------------------------------------------------------------------------
# ----------------------------- clean factor -----------------------------------
df_long$tr_ordered <- factor(
        df_long$tr_num,
        levels = c(paste0("tr", 1:TR_length))
    )
df_long$cue_ordered <- factor(
        df_long$cue,
        levels = c("cueH", "cueL")
    )
df_long$stim_ordered <- factor(
        df_long$stim,
        levels = c("stimH", "stimM", "stimL")
    )
df_long$sixcond <- factor(
        df_long$condition,
        levels = c("cueH_stimH", "cueL_stimH",
                   "cueH_stimM", "cueL_stimM",
                   "cueH_stimL", "cueL_stimL")
)

# Add a numeric TR variable for proper sorting
df_long$tr_numeric <- as.numeric(gsub("tr", "", df_long$tr_num))

# Extract numeric run value
df_long$run_numeric <- as.numeric(gsub("run-", "", df_long$run))

# For each subject and session, determine which run came first (1) vs second (2)
df_long <- df_long %>%
    group_by(sub, ses) %>%
    mutate(run_within_session = rank(run_numeric)) %>%  # 1 = first run, 2 = second run
    ungroup()

# Create global run order: session order (1,3,4) * 2 runs each = 6 runs total
df_long <- df_long %>%
    mutate(
        session_order = case_when(
            ses == "ses-01" ~ 1,
            ses == "ses-03" ~ 2,
            ses == "ses-04" ~ 3,
            TRUE ~ NA_real_
        ),
        run_order_global = as.integer((session_order - 1) * 2 + run_within_session)  # FORCE INTEGER
    )

# Filter for only high intensity stimuli (cueH_stimH and cueL_stimH)
df_high <- df_long %>%
    filter(sixcond %in% c("cueH_stimH", "cueL_stimH"))  # FIXED: was cueH_stimL

# --------------------------- summary statistics BY RUN ORDER -------------------
# Calculate subject-wise means for each GLOBAL RUN ORDER (not by session!)
subjectwise_run <- meanSummary(df_high,
                               c("sub", "run_order_global", "tr_ordered", "tr_numeric", "sixcond"), 
                               "tr_value")

# Calculate group-wise statistics - only group by run_order_global, NOT by session
groupwise_run <- summarySEwithin(
    data = subjectwise_run,
    measurevar = "mean_per_sub",
    withinvars = c("sixcond", "tr_ordered", "tr_numeric", "run_order_global"),
    idvar = "sub"
)
groupwise_run$task <- taskname

# Sort by run order and TR
groupwise_run <- groupwise_run %>%
    arrange(run_order_global, sixcond, tr_numeric)

# --------------------------------- plot each run ------------------------------
LINEIV1 = "tr_ordered"
LINEIV2 = "sixcond"
MEAN = "mean_per_sub_norm_mean"
ERROR = "se"

# Create a list to store plots - USE STANDARD LIST
plot_list <- list()

# Get unique runs in order
run_orders <- sort(unique(groupwise_run$run_order_global))

# Create labels for each run
run_labels <- c(
    "Run 1: Session 1, 1st run",
    "Run 2: Session 1, 2nd run",
    "Run 3: Session 3, 1st run",
    "Run 4: Session 3, 2nd run",
    "Run 5: Session 4, 1st run",
    "Run 6: Session 4, 2nd run"
)

for (i in run_orders) {
    # Subset data for this run
    run_data <- groupwise_run %>% 
        filter(run_order_global == i) %>%
        arrange(sixcond, tr_numeric)
    
    # Recreate tr_ordered with ONLY the levels present, in correct order
    unique_trs <- run_data %>%
        distinct(tr_ordered, tr_numeric) %>%
        arrange(tr_numeric) %>%
        pull(tr_ordered)
    
    run_data$tr_ordered <- factor(run_data$tr_ordered, levels = unique_trs)
    
    # Create plot
    p <- plot_timeseries_bar(
        run_data,
        LINEIV1, LINEIV2, MEAN, ERROR,
        xlab = "Time (seconds)",
        ylab = "Brain Activity (AU)",
        ggtitle = paste0(run_labels[i], " (N = ", unique(run_data$N), ")"),
        color = c("red", "#5f0f40")
    )
    
    # Create time labels
    n_trs <- length(unique_trs)
    time_points <- seq(0, (n_trs - 1) * 0.46, 0.46)
    time_labels <- round(time_points, 2)
    
    p <- p + 
        scale_x_discrete(labels = time_labels) +
        theme_classic() +
        theme(legend.position = "bottom")
    
    # Store with NUMERIC index, not the value of i
    plot_list[[length(plot_list) + 1]] <- p
    
    # Print the plot
    print(p)
}

# Create a combined figure with all 6 runs
library(gridExtra)
combined_plot <- grid.arrange(grobs = plot_list, ncol = 2, nrow = 3)

# To view the combined plot
print(combined_plot)
```

```{r}
roi_list <- c('rINS', 'TPJ', 'dACC', 'PHG', 'V1', 'SM', 'MT', 'RSC', 'LOC', 'FFC', 'PIT', 'pSTS', 'AIP', 'premotor')
run_types <- c("pain", "vicarious", "cognitive")
TR_length <- 42

for (ROI in roi_list) {
    datadir = file.path(main_dir, "analysis/fmri/spm/fir/ttl2par")
    taskname = 'pain'
    exclude <- "sub-0001"
    filename <- paste0("sub-*", "*roi-", ROI, "_tr-42.csv")
    common_path <- Sys.glob(file.path(datadir, "sub-*", filename))
    filter_path <- common_path[!str_detect(common_path, pattern = exclude)]
    
    df <- do.call("rbind.fill", lapply(filter_path, FUN = function(files) {
        read.table(files, header = TRUE, sep = ",")
    }))
    
    for (run_type in run_types) {
        print(paste0("Processing ROI: ", ROI, " - Run type: ", run_type))
        
        filtered_df <- df[!(df$condition == "rating" | df$condition == "cue" | df$runtype != run_type), ]
        parsed_df <- filtered_df %>%
            separate(condition, into = c("cue", "stim"), sep = "_", remove = FALSE)
        
        # --------------------- subset regions based on ROI ----------------------------
        df_long <- pivot_longer(parsed_df, cols = starts_with("tr"), names_to = "tr_num", values_to = "tr_value")
        
        # ===============================================================================
        # START: RUN ORDER ANALYSIS
        # ===============================================================================
        
        # ----------------------------- clean factor -----------------------------------
        df_long$tr_ordered <- factor(
            df_long$tr_num,
            levels = c(paste0("tr", 1:TR_length))
        )
        df_long$cue_ordered <- factor(
            df_long$cue,
            levels = c("cueH", "cueL")
        )
        df_long$stim_ordered <- factor(
            df_long$stim,
            levels = c("stimH", "stimM", "stimL")
        )
        df_long$sixcond <- factor(
            df_long$condition,
            levels = c("cueH_stimH", "cueL_stimH",
                       "cueH_stimM", "cueL_stimM",
                       "cueH_stimL", "cueL_stimL")
        )
        
        # Add a numeric TR variable for proper sorting
        df_long$tr_numeric <- as.numeric(gsub("tr", "", df_long$tr_num))
        
        # Extract numeric run value
        df_long$run_numeric <- as.numeric(gsub("run-", "", df_long$run))
        
        # For each subject and session, determine which run came first (1) vs second (2)
        df_long <- df_long %>%
            group_by(sub, ses) %>%
            mutate(run_within_session = rank(run_numeric)) %>%  # 1 = first run, 2 = second run
            ungroup()
        
        # Create global run order: session order (1,3,4) * 2 runs each = 6 runs total
        df_long <- df_long %>%
            mutate(
                session_order = case_when(
                    ses == "ses-01" ~ 1,
                    ses == "ses-03" ~ 2,
                    ses == "ses-04" ~ 3,
                    TRUE ~ NA_real_
                ),
                run_order_global = as.integer((session_order - 1) * 2 + run_within_session)
            )
        
        # Filter for only high intensity stimuli (cueH_stimH and cueL_stimH)
        df_high <- df_long %>%
            filter(sixcond %in% c("cueH_stimH", "cueL_stimH"))
        
        # --------------------------- summary statistics BY RUN ORDER -------------------
        # Calculate subject-wise means for each GLOBAL RUN ORDER (not by session!)
        subjectwise_run <- meanSummary(df_high,
                                       c("sub", "run_order_global", "tr_ordered", "tr_numeric", "sixcond"), 
                                       "tr_value")
        
        # Calculate group-wise statistics - only group by run_order_global, NOT by session
        groupwise_run <- summarySEwithin(
            data = subjectwise_run,
            measurevar = "mean_per_sub",
            withinvars = c("sixcond", "tr_ordered", "tr_numeric", "run_order_global"),
            idvar = "sub"
        )
        groupwise_run$task <- taskname
        
        # Sort by run order and TR
        groupwise_run <- groupwise_run %>%
            arrange(run_order_global, sixcond, tr_numeric)
        
        # --------------------------------- plot each run ------------------------------
        LINEIV1 = "tr_ordered"
        LINEIV2 = "sixcond"
        MEAN = "mean_per_sub_norm_mean"
        ERROR = "se"
        
        # Create a list to store plots
        plot_list <- list()
        
        # Get unique runs in order
        run_orders <- sort(unique(groupwise_run$run_order_global))
        
        # Create labels for each run
        run_labels <- c(
            "Run 1: Session 1, 1st run",
            "Run 2: Session 1, 2nd run",
            "Run 3: Session 3, 1st run",
            "Run 4: Session 3, 2nd run",
            "Run 5: Session 4, 1st run",
            "Run 6: Session 4, 2nd run"
        )
        
        for (i in run_orders) {
            # Subset data for this run
            run_data <- groupwise_run %>% 
                filter(run_order_global == i) %>%
                arrange(sixcond, tr_numeric)
            
            # Recreate tr_ordered with ONLY the levels present, in correct order
            unique_trs <- run_data %>%
                distinct(tr_ordered, tr_numeric) %>%
                arrange(tr_numeric) %>%
                pull(tr_ordered)
            
            run_data$tr_ordered <- factor(run_data$tr_ordered, levels = unique_trs)
            
            # Create plot
            p <- plot_timeseries_bar(
                run_data,
                LINEIV1, LINEIV2, MEAN, ERROR,
                xlab = "Time (seconds)",
                ylab = "Brain Activity (AU)",
                ggtitle = paste0(ROI, " - ", run_type, " - ", run_labels[i], " (N = ", unique(run_data$N), ")"),
                color = c("red", "#5f0f40")
            )
            
            # Create time labels
            n_trs <- length(unique_trs)
            time_points <- seq(0, (n_trs - 1) * 0.46, 0.46)
            time_labels <- round(time_points, 2)
            
            p <- p + 
                scale_x_discrete(labels = time_labels) +
                theme_classic() +
                theme(legend.position = "bottom")
            
            # Store plot
            plot_list[[length(plot_list) + 1]] <- p
            
            # Print the plot
            print(p)
        }
        
        # Create a combined figure with all 6 runs
        library(gridExtra)
        combined_plot <- grid.arrange(
            grobs = plot_list, 
            ncol = 2, 
            nrow = 3,
            top = textGrob(paste0("ROI: ", ROI, " - ", run_type, " - All Runs"), 
                           gp = gpar(fontsize = 16, fontface = "bold"))
        )
        
        # Print the combined plot
        print(combined_plot)
        
        # Optional: Save the combined plot
        ggsave(
            filename = paste0("roi_", ROI, "_", run_type, "_all_runs.png"),
            plot = combined_plot,
            width = 12,
            height = 14,
            dpi = 300
        )
        
        # ===============================================================================
        # END: RUN ORDER ANALYSIS
        # ===============================================================================
    }
}
```
