==> scripts/step02_R_bookdown/93_iv-cue_dv-sclbeta.Rmd <==
# [physio] SCL {#ch93_SCL}

---

title: "93_iv-cue_dv-sclbeta"
output: html_document
date: "2023-07-20"

---

## Outline

## load data

## subjectwise, groupwise mean

```{r message=TRUE, warning=TRUE, include=FALSE}
library(car)
library(psych)
library(reshape)
library(PupillometryR)
library(plyr); library(dplyr)
library(tidyselect)
library(tidyr)
library(stringr)
library(lmerTest)
library(gghalves)
source("https://gist.githubusercontent.com/benmarwick/2a1bb0133ff568cbe28d/raw/fb53bd97121f7f9ce947837ef1a4c65a73bffb3f/geom_flat_violin.R")
main_dir = dirname(dirname(getwd()))
file.sources = list.files(file.path(main_dir, "scripts/step02_R/utils"),
                          pattern="*.R",
                          full.names=TRUE,
                          ignore.case=TRUE)
sapply(file.sources,source,.GlobalEnv)
```

```{r function::simple_contrasts_singletrial, message=FALSE, warning=FALSE, include=FALSE, paged.print=FALSE}
simple_contrasts_singletrial <- function(df) {
# [ CONTRASTS ]  ________________________________________________________________________________ # nolint

df$cue_factor <- factor(df$cue_type)

df$cue_name[df$cue_type == "intercept"] <- "intercept"
df$cue_name[df$cue_type == "low_cue"] <- "low"
df$cue_name[df$cue_type == "high_cue"] <- "high"


df$cue_ordered <- factor(
        df$cue_name,
        levels = c("intercept", "low", "high")
    )

return(df)
}
```

```{r}
beta <- read.table(file = "/Volumes/spacetop_projects_cue/analysis/physio/glm/pmod-cue/glm-pmodcue_task-pain_scr.tsv", sep = '\t', header = TRUE)

beta_long <- gather(beta, key = "cue_type", value = "scl_value", intercept, low_cue, high_cue)
beta_con <- simple_contrasts_singletrial(beta_long)

```

```{r}

# ----------------------------------------------------------------------
#                     summary statistics for plots
# ----------------------------------------------------------------------
subject <- "sub"
model_iv <- "cue_ordered"
dv <- "scl_value"
dv_keyword <- "sclbeta"
taskname <- "pain"
analysis_dir <- file.path(main_dir, "analysis", "mixedeffect", "model93_physio", as.character(Sys.Date()))

# ======= NOTE: calculate mean and se ----------------------------------
SCLcue_subjectwise <- meanSummary(beta_con,
                                      c(subject, model_iv), dv)
SCLcue_groupwise <- summarySEwithin(
  data = SCLcue_subjectwise,
  measurevar = "mean_per_sub",
  withinvars = c(model_iv),
  idvar = subject
)
SCLcue_groupwise$task <- taskname
# https://stackoverflow.com/questions/29402528/append-data-frames-together-in-a-for-loop/29419402

# combined_se_calc_cooksd <-NPSstimcue_groupwise
# calculate mean and se
sub_mean <- "mean_per_sub"
group_mean <- "mean_per_sub_norm_mean"
se <- "se"
subject <- "sub"
ggtitle <- paste(taskname, " - ", dv, "Cooksd removed")
title <- paste(taskname, " - ", dv)
xlab <- ""
ylab <- "ROI average activation (A.U)"
ylim <- c(-10, 60)

if (any(startsWith(dv_keyword, c("expect", "Expect")))) {
  color <- c("#1B9E77", "#D95F02")
} else {
  color <- c("#4274AD", "#C5263A")
} # if keyword starts with
plot_savefname <- file.path(
  analysis_dir,
  paste(
    "raincloud_task-", taskname, "_rating-", dv_keyword, "_", as.character(Sys.Date()), "_cooksd.png",
    sep = ""
  )
)

# ----------------------------------------------------------------------
#                            raincloudplots
# ----------------------------------------------------------------------
# combined_se_calc_cooksd <-NPSstimcue_groupwise
# calculate mean and se
sub_mean <- "mean_per_sub"
group_mean <- "mean_per_sub_norm_mean"
se <- "se"
subject <- "sub"
ggtitle <- paste( dv)
title <- paste( dv)
xlab <- ""
ylab <- "SCL beta coefficients (A.U.)"

dv_keyword <- "fdmean"
if (any(startsWith(dv_keyword, c("expect", "Expect")))) {
  color <- c("gray", "blue",  "red")
} else {
  color <- c("gray", "blue", "red")
} # if keyword starts with
plot_savefname <- file.path(
  analysis_dir,
  paste(
    "raincloud_qc-", dv_keyword, "_", as.character(Sys.Date()), "_cooksd.png",
    sep = ""
  )
)

# ----------------------------------------------------------------------
#                            raincloudplots
# ----------------------------------------------------------------------


ylim <- c(-1, 1.5)
# taskname = "pain"
w <- 10; h <- 5
g <- plot_halfrainclouds_onefactor(
  SCLcue_subjectwise,
  SCLcue_groupwise,
  model_iv,
  sub_mean,
  group_mean,
  se,
  subject,
  ggtitle,
  title,
  xlab,
  ylab,
  taskname,
  ylim,
  w,
  h,
  dv_keyword,
  color,
  plot_savefname
)
g <- g + theme_bw() + theme_classic()
print(g)
SCLcue_groupwise$task = taskname

k <- plot_lineplot_onefactorthick(SCLcue_groupwise,
                             taskname = "pain",
                        iv = "cue_ordered",
                        mean = "mean_per_sub_norm_mean", error = "se",
                        color = c("intercept" = "gray",
                                  "high" = "red",
                                  "low" = "blue"),
                        ggtitle = title,
                        xlab = "Cue level", ylab = "SCL activation (A.U.)")
k <- k + theme(aspect.ratio=.8)
print(k)
```
